Struct aws_sdk_glue::client::fluent_builders::CreateMLTransform

pub struct CreateMLTransform<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateMLTransform<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateMlTransformOutput, SdkError<CreateMLTransformError>> where
R::Policy: SmithyRetryPolicy<CreateMlTransformInputOperationOutputAlias, CreateMlTransformOutput, CreateMLTransformError, CreateMlTransformInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

The unique name that you give the transform when you create it.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the machine learning transform that is being defined. The default is an empty string.

pub fn set_description(self, input: Option<String>) -> Self

pub fn input_record_tables(self, inp: impl Into<GlueTable>) -> Self

Appends an item to InputRecordTables.

To override the contents of this collection use set_input_record_tables.

A list of Glue table definitions used by the transform.

pub fn set_input_record_tables(self, input: Option<Vec<GlueTable>>) -> Self

pub fn parameters(self, inp: TransformParameters) -> Self

The algorithmic parameters that are specific to the transform type used. Conditionally dependent on the transform type.

pub fn set_parameters(self, input: Option<TransformParameters>) -> Self

pub fn role(self, inp: impl Into<String>) -> Self

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions. The required permissions include both Glue service role permissions to Glue resources, and Amazon S3 permissions required by the transform.

This role needs Glue service role permissions to allow access to resources in Glue. See Attach a Policy to IAM Users That Access Glue.

This role needs permission to your Amazon Simple Storage Service (Amazon S3) sources, targets, temporary directory, scripts, and any libraries used by the task run for this transform.

pub fn set_role(self, input: Option<String>) -> Self

pub fn glue_version(self, inp: impl Into<String>) -> Self

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

pub fn set_glue_version(self, input: Option<String>) -> Self

pub fn max_capacity(self, inp: f64) -> Self

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

pub fn set_max_capacity(self, input: Option<f64>) -> Self

pub fn worker_type(self, inp: WorkerType) -> Self

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

pub fn set_worker_type(self, input: Option<WorkerType>) -> Self

pub fn number_of_workers(self, inp: i32) -> Self

The number of workers of a defined workerType that are allocated when this task runs.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

pub fn set_number_of_workers(self, input: Option<i32>) -> Self

pub fn timeout(self, inp: i32) -> Self

The timeout of the task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

pub fn set_timeout(self, input: Option<i32>) -> Self

pub fn max_retries(self, inp: i32) -> Self

The maximum number of times to retry a task for this transform after a task run fails.

pub fn set_max_retries(self, input: Option<i32>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

The tags to use with this machine learning transform. You may use tags to limit access to the machine learning transform. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn transform_encryption(self, inp: TransformEncryption) -> Self

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

pub fn set_transform_encryption(
self,
input: Option<TransformEncryption>
) -> Self