Struct aws_sdk_glue::client::fluent_builders::CreateSchema

pub struct CreateSchema<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateSchema<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateSchemaOutput, SdkError<CreateSchemaError>> where
R::Policy: SmithyRetryPolicy<CreateSchemaInputOperationOutputAlias, CreateSchemaOutput, CreateSchemaError, CreateSchemaInputOperationRetryAlias>,

pub fn registry_id(self, inp: RegistryId) -> Self

This is a wrapper shape to contain the registry identity fields. If this is not provided, the default registry will be used. The ARN format for the same will be: arn:aws:glue:us-east-2::registry/default-registry:random-5-letter-id.

pub fn set_registry_id(self, input: Option<RegistryId>) -> Self

pub fn schema_name(self, inp: impl Into<String>) -> Self

Name of the schema to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

pub fn set_schema_name(self, input: Option<String>) -> Self

pub fn data_format(self, inp: DataFormat) -> Self

The data format of the schema definition. Currently AVRO and JSON are supported.

pub fn set_data_format(self, input: Option<DataFormat>) -> Self

pub fn compatibility(self, inp: Compatibility) -> Self

The compatibility mode of the schema. The possible values are:

NONE: No compatibility mode applies. You can use this choice in development scenarios or if you do not know the compatibility mode that you want to apply to schemas. Any new version added will be accepted without undergoing a compatibility check.

DISABLED: This compatibility choice prevents versioning for a particular schema. You can use this choice to prevent future versioning of a schema.

BACKWARD: This compatibility choice is recommended as it allows data receivers to read both the current and one previous schema version. This means that for instance, a new schema version cannot drop data fields or change the type of these fields, so they can't be read by readers using the previous version.

BACKWARD_ALL: This compatibility choice allows data receivers to read both the current and all previous schema versions. You can use this choice when you need to delete fields or add optional fields, and check compatibility against all previous schema versions.

FORWARD: This compatibility choice allows data receivers to read both the current and one next schema version, but not necessarily later versions. You can use this choice when you need to add fields or delete optional fields, but only check compatibility against the last schema version.

FORWARD_ALL: This compatibility choice allows data receivers to read written by producers of any new registered schema. You can use this choice when you need to add fields or delete optional fields, and check compatibility against all previous schema versions.

FULL: This compatibility choice allows data receivers to read data written by producers using the previous or next version of the schema, but not necessarily earlier or later versions. You can use this choice when you need to add or remove optional fields, but only check compatibility against the last schema version.

FULL_ALL: This compatibility choice allows data receivers to read data written by producers using all previous schema versions. You can use this choice when you need to add or remove optional fields, and check compatibility against all previous schema versions.

pub fn set_compatibility(self, input: Option<Compatibility>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

An optional description of the schema. If description is not provided, there will not be any automatic default value for this.

pub fn set_description(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API. If specified, follows the Amazon Web Services tags-on-create pattern.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn schema_definition(self, inp: impl Into<String>) -> Self

The schema definition using the DataFormat setting for SchemaName.

pub fn set_schema_definition(self, input: Option<String>) -> Self