Struct aws_sdk_glue::client::fluent_builders::StartJobRun

pub struct StartJobRun<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> StartJobRun<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<StartJobRunOutput, SdkError<StartJobRunError>> where
R::Policy: SmithyRetryPolicy<StartJobRunInputOperationOutputAlias, StartJobRunOutput, StartJobRunError, StartJobRunInputOperationRetryAlias>,

pub fn job_name(self, inp: impl Into<String>) -> Self

The name of the job definition to use.

pub fn set_job_name(self, input: Option<String>) -> Self

pub fn job_run_id(self, inp: impl Into<String>) -> Self

The ID of a previous JobRun to retry.

pub fn set_job_run_id(self, input: Option<String>) -> Self

pub fn arguments(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Arguments.

To override the contents of this collection use set_arguments.

The job arguments specifically for this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

pub fn set_arguments(self, input: Option<HashMap<String, String>>) -> Self

pub fn allocated_capacity(self, inp: i32) -> Self

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

pub fn set_allocated_capacity(self, input: Option<i32>) -> Self

pub fn timeout(self, inp: i32) -> Self

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

pub fn set_timeout(self, input: Option<i32>) -> Self

pub fn max_capacity(self, inp: f64) -> Self

The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job, or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

pub fn set_max_capacity(self, input: Option<f64>) -> Self

pub fn security_configuration(self, inp: impl Into<String>) -> Self

The name of the SecurityConfiguration structure to be used with this job run.

pub fn set_security_configuration(self, input: Option<String>) -> Self

pub fn notification_property(self, inp: NotificationProperty) -> Self

Specifies configuration properties of a job run notification.

pub fn set_notification_property(
self,
input: Option<NotificationProperty>
) -> Self

pub fn worker_type(self, inp: WorkerType) -> Self

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

pub fn set_worker_type(self, input: Option<WorkerType>) -> Self

pub fn number_of_workers(self, inp: i32) -> Self

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

pub fn set_number_of_workers(self, input: Option<i32>) -> Self