Struct aws_sdk_glue::client::fluent_builders::UpdateDevEndpoint

pub struct UpdateDevEndpoint<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateDevEndpoint<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateDevEndpointOutput, SdkError<UpdateDevEndpointError>> where
R::Policy: SmithyRetryPolicy<UpdateDevEndpointInputOperationOutputAlias, UpdateDevEndpointOutput, UpdateDevEndpointError, UpdateDevEndpointInputOperationRetryAlias>,

pub fn endpoint_name(self, inp: impl Into<String>) -> Self

The name of the DevEndpoint to be updated.

pub fn set_endpoint_name(self, input: Option<String>) -> Self

pub fn public_key(self, inp: impl Into<String>) -> Self

The public key for the DevEndpoint to use.

pub fn set_public_key(self, input: Option<String>) -> Self

pub fn add_public_keys(self, inp: impl Into<String>) -> Self

Appends an item to AddPublicKeys.

To override the contents of this collection use set_add_public_keys.

The list of public keys for the DevEndpoint to use.

pub fn set_add_public_keys(self, input: Option<Vec<String>>) -> Self

pub fn delete_public_keys(self, inp: impl Into<String>) -> Self

Appends an item to DeletePublicKeys.

To override the contents of this collection use set_delete_public_keys.

The list of public keys to be deleted from the DevEndpoint.

pub fn set_delete_public_keys(self, input: Option<Vec<String>>) -> Self

pub fn custom_libraries(self, inp: DevEndpointCustomLibraries) -> Self

Custom Python or Java libraries to be loaded in the DevEndpoint.

pub fn set_custom_libraries(
self,
input: Option<DevEndpointCustomLibraries>
) -> Self

pub fn update_etl_libraries(self, inp: bool) -> Self

True if the list of custom libraries to be loaded in the development endpoint needs to be updated, or False if otherwise.

pub fn set_update_etl_libraries(self, input: Option<bool>) -> Self

pub fn delete_arguments(self, inp: impl Into<String>) -> Self

Appends an item to DeleteArguments.

To override the contents of this collection use set_delete_arguments.

The list of argument keys to be deleted from the map of arguments used to configure the DevEndpoint.

pub fn set_delete_arguments(self, input: Option<Vec<String>>) -> Self

pub fn add_arguments(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to AddArguments.

To override the contents of this collection use set_add_arguments.

The map of arguments to add the map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

pub fn set_add_arguments(self, input: Option<HashMap<String, String>>) -> Self