Struct aws_sdk_glue::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSGlue.

This client allows ergonomic access to a AWSGlue-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_create_partition(&self) -> BatchCreatePartition<C, M, R>

pub fn batch_delete_connection(&self) -> BatchDeleteConnection<C, M, R>

pub fn batch_delete_partition(&self) -> BatchDeletePartition<C, M, R>

pub fn batch_delete_table(&self) -> BatchDeleteTable<C, M, R>

pub fn batch_delete_table_version(&self) -> BatchDeleteTableVersion<C, M, R>

pub fn batch_get_blueprints(&self) -> BatchGetBlueprints<C, M, R>

pub fn batch_get_crawlers(&self) -> BatchGetCrawlers<C, M, R>

pub fn batch_get_dev_endpoints(&self) -> BatchGetDevEndpoints<C, M, R>

pub fn batch_get_jobs(&self) -> BatchGetJobs<C, M, R>

pub fn batch_get_partition(&self) -> BatchGetPartition<C, M, R>

pub fn batch_get_triggers(&self) -> BatchGetTriggers<C, M, R>

pub fn batch_get_workflows(&self) -> BatchGetWorkflows<C, M, R>

pub fn batch_stop_job_run(&self) -> BatchStopJobRun<C, M, R>

pub fn batch_update_partition(&self) -> BatchUpdatePartition<C, M, R>

pub fn cancel_ml_task_run(&self) -> CancelMLTaskRun<C, M, R>

pub fn check_schema_version_validity(
&self
) -> CheckSchemaVersionValidity<C, M, R>

pub fn create_blueprint(&self) -> CreateBlueprint<C, M, R>

pub fn create_classifier(&self) -> CreateClassifier<C, M, R>

pub fn create_connection(&self) -> CreateConnection<C, M, R>

pub fn create_crawler(&self) -> CreateCrawler<C, M, R>

pub fn create_database(&self) -> CreateDatabase<C, M, R>

pub fn create_dev_endpoint(&self) -> CreateDevEndpoint<C, M, R>

pub fn create_job(&self) -> CreateJob<C, M, R>

pub fn create_ml_transform(&self) -> CreateMLTransform<C, M, R>

pub fn create_partition(&self) -> CreatePartition<C, M, R>

pub fn create_partition_index(&self) -> CreatePartitionIndex<C, M, R>

pub fn create_registry(&self) -> CreateRegistry<C, M, R>

pub fn create_schema(&self) -> CreateSchema<C, M, R>

pub fn create_script(&self) -> CreateScript<C, M, R>

pub fn create_security_configuration(
&self
) -> CreateSecurityConfiguration<C, M, R>

pub fn create_table(&self) -> CreateTable<C, M, R>

pub fn create_trigger(&self) -> CreateTrigger<C, M, R>

pub fn create_user_defined_function(&self) -> CreateUserDefinedFunction<C, M, R>

pub fn create_workflow(&self) -> CreateWorkflow<C, M, R>

pub fn delete_blueprint(&self) -> DeleteBlueprint<C, M, R>

pub fn delete_classifier(&self) -> DeleteClassifier<C, M, R>

pub fn delete_column_statistics_for_partition(
&self
) -> DeleteColumnStatisticsForPartition<C, M, R>

pub fn delete_column_statistics_for_table(
&self
) -> DeleteColumnStatisticsForTable<C, M, R>

pub fn delete_connection(&self) -> DeleteConnection<C, M, R>

pub fn delete_crawler(&self) -> DeleteCrawler<C, M, R>

pub fn delete_database(&self) -> DeleteDatabase<C, M, R>

pub fn delete_dev_endpoint(&self) -> DeleteDevEndpoint<C, M, R>

pub fn delete_job(&self) -> DeleteJob<C, M, R>

pub fn delete_ml_transform(&self) -> DeleteMLTransform<C, M, R>

pub fn delete_partition(&self) -> DeletePartition<C, M, R>

pub fn delete_partition_index(&self) -> DeletePartitionIndex<C, M, R>

pub fn delete_registry(&self) -> DeleteRegistry<C, M, R>

pub fn delete_resource_policy(&self) -> DeleteResourcePolicy<C, M, R>

pub fn delete_schema(&self) -> DeleteSchema<C, M, R>

pub fn delete_schema_versions(&self) -> DeleteSchemaVersions<C, M, R>

pub fn delete_security_configuration(
&self
) -> DeleteSecurityConfiguration<C, M, R>

pub fn delete_table(&self) -> DeleteTable<C, M, R>

pub fn delete_table_version(&self) -> DeleteTableVersion<C, M, R>

pub fn delete_trigger(&self) -> DeleteTrigger<C, M, R>

pub fn delete_user_defined_function(&self) -> DeleteUserDefinedFunction<C, M, R>

pub fn delete_workflow(&self) -> DeleteWorkflow<C, M, R>

pub fn get_blueprint(&self) -> GetBlueprint<C, M, R>

pub fn get_blueprint_run(&self) -> GetBlueprintRun<C, M, R>

pub fn get_blueprint_runs(&self) -> GetBlueprintRuns<C, M, R>

pub fn get_catalog_import_status(&self) -> GetCatalogImportStatus<C, M, R>

pub fn get_classifier(&self) -> GetClassifier<C, M, R>

pub fn get_classifiers(&self) -> GetClassifiers<C, M, R>

pub fn get_column_statistics_for_partition(
&self
) -> GetColumnStatisticsForPartition<C, M, R>

pub fn get_column_statistics_for_table(
&self
) -> GetColumnStatisticsForTable<C, M, R>

pub fn get_connection(&self) -> GetConnection<C, M, R>

pub fn get_connections(&self) -> GetConnections<C, M, R>

pub fn get_crawler(&self) -> GetCrawler<C, M, R>

pub fn get_crawler_metrics(&self) -> GetCrawlerMetrics<C, M, R>

pub fn get_crawlers(&self) -> GetCrawlers<C, M, R>

pub fn get_database(&self) -> GetDatabase<C, M, R>

pub fn get_databases(&self) -> GetDatabases<C, M, R>

pub fn get_data_catalog_encryption_settings(
&self
) -> GetDataCatalogEncryptionSettings<C, M, R>

pub fn get_dataflow_graph(&self) -> GetDataflowGraph<C, M, R>

pub fn get_dev_endpoint(&self) -> GetDevEndpoint<C, M, R>

pub fn get_dev_endpoints(&self) -> GetDevEndpoints<C, M, R>

pub fn get_job(&self) -> GetJob<C, M, R>

pub fn get_job_bookmark(&self) -> GetJobBookmark<C, M, R>

pub fn get_job_run(&self) -> GetJobRun<C, M, R>

pub fn get_job_runs(&self) -> GetJobRuns<C, M, R>

pub fn get_jobs(&self) -> GetJobs<C, M, R>

pub fn get_mapping(&self) -> GetMapping<C, M, R>

pub fn get_ml_task_run(&self) -> GetMLTaskRun<C, M, R>

pub fn get_ml_task_runs(&self) -> GetMLTaskRuns<C, M, R>

pub fn get_ml_transform(&self) -> GetMLTransform<C, M, R>

pub fn get_ml_transforms(&self) -> GetMLTransforms<C, M, R>

pub fn get_partition(&self) -> GetPartition<C, M, R>

pub fn get_partition_indexes(&self) -> GetPartitionIndexes<C, M, R>

pub fn get_partitions(&self) -> GetPartitions<C, M, R>

pub fn get_plan(&self) -> GetPlan<C, M, R>

pub fn get_registry(&self) -> GetRegistry<C, M, R>

pub fn get_resource_policies(&self) -> GetResourcePolicies<C, M, R>

pub fn get_resource_policy(&self) -> GetResourcePolicy<C, M, R>

pub fn get_schema(&self) -> GetSchema<C, M, R>

pub fn get_schema_by_definition(&self) -> GetSchemaByDefinition<C, M, R>

pub fn get_schema_version(&self) -> GetSchemaVersion<C, M, R>

pub fn get_schema_versions_diff(&self) -> GetSchemaVersionsDiff<C, M, R>

pub fn get_security_configuration(&self) -> GetSecurityConfiguration<C, M, R>

pub fn get_security_configurations(&self) -> GetSecurityConfigurations<C, M, R>

pub fn get_table(&self) -> GetTable<C, M, R>

pub fn get_tables(&self) -> GetTables<C, M, R>

pub fn get_table_version(&self) -> GetTableVersion<C, M, R>

pub fn get_table_versions(&self) -> GetTableVersions<C, M, R>

pub fn get_tags(&self) -> GetTags<C, M, R>

pub fn get_trigger(&self) -> GetTrigger<C, M, R>

pub fn get_triggers(&self) -> GetTriggers<C, M, R>

pub fn get_user_defined_function(&self) -> GetUserDefinedFunction<C, M, R>

pub fn get_user_defined_functions(&self) -> GetUserDefinedFunctions<C, M, R>

pub fn get_workflow(&self) -> GetWorkflow<C, M, R>

pub fn get_workflow_run(&self) -> GetWorkflowRun<C, M, R>

pub fn get_workflow_run_properties(&self) -> GetWorkflowRunProperties<C, M, R>

pub fn get_workflow_runs(&self) -> GetWorkflowRuns<C, M, R>

pub fn import_catalog_to_glue(&self) -> ImportCatalogToGlue<C, M, R>

pub fn list_blueprints(&self) -> ListBlueprints<C, M, R>

pub fn list_crawlers(&self) -> ListCrawlers<C, M, R>

pub fn list_dev_endpoints(&self) -> ListDevEndpoints<C, M, R>

pub fn list_jobs(&self) -> ListJobs<C, M, R>

pub fn list_ml_transforms(&self) -> ListMLTransforms<C, M, R>

pub fn list_registries(&self) -> ListRegistries<C, M, R>

pub fn list_schemas(&self) -> ListSchemas<C, M, R>

pub fn list_schema_versions(&self) -> ListSchemaVersions<C, M, R>

pub fn list_triggers(&self) -> ListTriggers<C, M, R>

pub fn list_workflows(&self) -> ListWorkflows<C, M, R>

pub fn put_data_catalog_encryption_settings(
&self
) -> PutDataCatalogEncryptionSettings<C, M, R>

pub fn put_resource_policy(&self) -> PutResourcePolicy<C, M, R>

pub fn put_schema_version_metadata(&self) -> PutSchemaVersionMetadata<C, M, R>

pub fn put_workflow_run_properties(&self) -> PutWorkflowRunProperties<C, M, R>

pub fn query_schema_version_metadata(
&self
) -> QuerySchemaVersionMetadata<C, M, R>

pub fn register_schema_version(&self) -> RegisterSchemaVersion<C, M, R>

pub fn remove_schema_version_metadata(
&self
) -> RemoveSchemaVersionMetadata<C, M, R>

pub fn reset_job_bookmark(&self) -> ResetJobBookmark<C, M, R>

pub fn resume_workflow_run(&self) -> ResumeWorkflowRun<C, M, R>

pub fn search_tables(&self) -> SearchTables<C, M, R>

pub fn start_blueprint_run(&self) -> StartBlueprintRun<C, M, R>

pub fn start_crawler(&self) -> StartCrawler<C, M, R>

pub fn start_crawler_schedule(&self) -> StartCrawlerSchedule<C, M, R>

pub fn start_export_labels_task_run(&self) -> StartExportLabelsTaskRun<C, M, R>

pub fn start_import_labels_task_run(&self) -> StartImportLabelsTaskRun<C, M, R>

pub fn start_job_run(&self) -> StartJobRun<C, M, R>

pub fn start_ml_evaluation_task_run(&self) -> StartMLEvaluationTaskRun<C, M, R>

pub fn start_ml_labeling_set_generation_task_run(
&self
) -> StartMLLabelingSetGenerationTaskRun<C, M, R>

pub fn start_trigger(&self) -> StartTrigger<C, M, R>

pub fn start_workflow_run(&self) -> StartWorkflowRun<C, M, R>

pub fn stop_crawler(&self) -> StopCrawler<C, M, R>

pub fn stop_crawler_schedule(&self) -> StopCrawlerSchedule<C, M, R>

pub fn stop_trigger(&self) -> StopTrigger<C, M, R>

pub fn stop_workflow_run(&self) -> StopWorkflowRun<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_blueprint(&self) -> UpdateBlueprint<C, M, R>

pub fn update_classifier(&self) -> UpdateClassifier<C, M, R>

pub fn update_column_statistics_for_partition(
&self
) -> UpdateColumnStatisticsForPartition<C, M, R>

pub fn update_column_statistics_for_table(
&self
) -> UpdateColumnStatisticsForTable<C, M, R>

pub fn update_connection(&self) -> UpdateConnection<C, M, R>

pub fn update_crawler(&self) -> UpdateCrawler<C, M, R>

pub fn update_crawler_schedule(&self) -> UpdateCrawlerSchedule<C, M, R>

pub fn update_database(&self) -> UpdateDatabase<C, M, R>

pub fn update_dev_endpoint(&self) -> UpdateDevEndpoint<C, M, R>

pub fn update_job(&self) -> UpdateJob<C, M, R>

pub fn update_ml_transform(&self) -> UpdateMLTransform<C, M, R>

pub fn update_partition(&self) -> UpdatePartition<C, M, R>

pub fn update_registry(&self) -> UpdateRegistry<C, M, R>

pub fn update_schema(&self) -> UpdateSchema<C, M, R>

pub fn update_table(&self) -> UpdateTable<C, M, R>

pub fn update_trigger(&self) -> UpdateTrigger<C, M, R>

pub fn update_user_defined_function(&self) -> UpdateUserDefinedFunction<C, M, R>

pub fn update_workflow(&self) -> UpdateWorkflow<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self