Module aws_sdk_glue::error

Modules

access_denied_exception

See AccessDeniedException

already_exists_exception

See AlreadyExistsException

concurrent_modification_exception

See ConcurrentModificationException

concurrent_runs_exceeded_exception

See ConcurrentRunsExceededException

condition_check_failure_exception

See ConditionCheckFailureException

conflict_exception

See ConflictException

crawler_not_running_exception

See CrawlerNotRunningException

crawler_running_exception

See CrawlerRunningException

crawler_stopping_exception

See CrawlerStoppingException

entity_not_found_exception

See EntityNotFoundException

glue_encryption_exception

See GlueEncryptionException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

illegal_blueprint_state_exception

See IllegalBlueprintStateException

illegal_workflow_state_exception

See IllegalWorkflowStateException

internal_service_exception

See InternalServiceException

invalid_input_exception

See InvalidInputException

ml_transform_not_ready_exception

See MlTransformNotReadyException

no_schedule_exception

See NoScheduleException

operation_timeout_exception

See OperationTimeoutException

resource_number_limit_exceeded_exception

See ResourceNumberLimitExceededException

scheduler_not_running_exception

See SchedulerNotRunningException

scheduler_running_exception

See SchedulerRunningException

scheduler_transitioning_exception

See SchedulerTransitioningException

validation_exception

See ValidationException

version_mismatch_exception

See VersionMismatchException

Structs

AccessDeniedException

Access to a resource was denied.

AlreadyExistsException

A resource to be created or added already exists.

BatchCreatePartitionError

BatchDeleteConnectionError

BatchDeletePartitionError

BatchDeleteTableError

BatchDeleteTableVersionError

BatchGetBlueprintsError

BatchGetCrawlersError

BatchGetDevEndpointsError

BatchGetJobsError

BatchGetPartitionError

BatchGetTriggersError

BatchGetWorkflowsError

BatchStopJobRunError

BatchUpdatePartitionError

CancelMLTaskRunError

CheckSchemaVersionValidityError

ConcurrentModificationException

Two processes are trying to modify a resource simultaneously.

ConcurrentRunsExceededException

Too many jobs are being run concurrently.

ConditionCheckFailureException

A specified condition was not satisfied.

ConflictException

The CreatePartitions API was called on a table that has indexes enabled.

CrawlerNotRunningException

The specified crawler is not running.

CrawlerRunningException

The operation cannot be performed because the crawler is already running.

CrawlerStoppingException

The specified crawler is stopping.

CreateBlueprintError

CreateClassifierError

CreateConnectionError

CreateCrawlerError

CreateDatabaseError

CreateDevEndpointError

CreateJobError

CreateMLTransformError

CreatePartitionError

CreatePartitionIndexError

CreateRegistryError

CreateSchemaError

CreateScriptError

CreateSecurityConfigurationError

CreateTableError

CreateTriggerError

CreateUserDefinedFunctionError

CreateWorkflowError

DeleteBlueprintError

DeleteClassifierError

DeleteColumnStatisticsForPartitionError

DeleteColumnStatisticsForTableError

DeleteConnectionError

DeleteCrawlerError

DeleteDatabaseError

DeleteDevEndpointError

DeleteJobError

DeleteMLTransformError

DeletePartitionError

DeletePartitionIndexError

DeleteRegistryError

DeleteResourcePolicyError

DeleteSchemaError

DeleteSchemaVersionsError

DeleteSecurityConfigurationError

DeleteTableError

DeleteTableVersionError

DeleteTriggerError

DeleteUserDefinedFunctionError

DeleteWorkflowError

EntityNotFoundException

A specified entity does not exist

GetBlueprintError

GetBlueprintRunError

GetBlueprintRunsError

GetCatalogImportStatusError

GetClassifierError

GetClassifiersError

GetColumnStatisticsForPartitionError

GetColumnStatisticsForTableError

GetConnectionError

GetConnectionsError

GetCrawlerError

GetCrawlerMetricsError

GetCrawlersError

GetDataCatalogEncryptionSettingsError

GetDatabaseError

GetDatabasesError

GetDataflowGraphError

GetDevEndpointError

GetDevEndpointsError

GetJobBookmarkError

GetJobError

GetJobRunError

GetJobRunsError

GetJobsError

GetMLTaskRunError

GetMLTaskRunsError

GetMLTransformError

GetMLTransformsError

GetMappingError

GetPartitionError

GetPartitionIndexesError

GetPartitionsError

GetPlanError

GetRegistryError

GetResourcePoliciesError

GetResourcePolicyError

GetSchemaByDefinitionError

GetSchemaError

GetSchemaVersionError

GetSchemaVersionsDiffError

GetSecurityConfigurationError

GetSecurityConfigurationsError

GetTableError

GetTableVersionError

GetTableVersionsError

GetTablesError

GetTagsError

GetTriggerError

GetTriggersError

GetUserDefinedFunctionError

GetUserDefinedFunctionsError

GetWorkflowError

GetWorkflowRunError

GetWorkflowRunPropertiesError

GetWorkflowRunsError

GlueEncryptionException

An encryption operation failed.

IdempotentParameterMismatchException

The same unique identifier was associated with two different records.

IllegalBlueprintStateException

IllegalWorkflowStateException

The workflow is in an invalid state to perform a requested operation.

ImportCatalogToGlueError

InternalServiceException

An internal service error occurred.

InvalidInputException

The input provided was not valid.

ListBlueprintsError

ListCrawlersError

ListDevEndpointsError

ListJobsError

ListMLTransformsError

ListRegistriesError

ListSchemaVersionsError

ListSchemasError

ListTriggersError

ListWorkflowsError

MlTransformNotReadyException

The machine learning transform is not ready to run.

NoScheduleException

There is no applicable schedule.

OperationTimeoutException

The operation timed out.

PutDataCatalogEncryptionSettingsError

PutResourcePolicyError

PutSchemaVersionMetadataError

PutWorkflowRunPropertiesError

QuerySchemaVersionMetadataError

RegisterSchemaVersionError

RemoveSchemaVersionMetadataError

ResetJobBookmarkError

ResourceNumberLimitExceededException

A resource numerical limit was exceeded.

ResumeWorkflowRunError

SchedulerNotRunningException

The specified scheduler is not running.

SchedulerRunningException

The specified scheduler is already running.

SchedulerTransitioningException

The specified scheduler is transitioning.

SearchTablesError

StartBlueprintRunError

StartCrawlerError

StartCrawlerScheduleError

StartExportLabelsTaskRunError

StartImportLabelsTaskRunError

StartJobRunError

StartMLEvaluationTaskRunError

StartMLLabelingSetGenerationTaskRunError

StartTriggerError

StartWorkflowRunError

StopCrawlerError

StopCrawlerScheduleError

StopTriggerError

StopWorkflowRunError

TagResourceError

UntagResourceError

UpdateBlueprintError

UpdateClassifierError

UpdateColumnStatisticsForPartitionError

UpdateColumnStatisticsForTableError

UpdateConnectionError

UpdateCrawlerError

UpdateCrawlerScheduleError

UpdateDatabaseError

UpdateDevEndpointError

UpdateJobError

UpdateMLTransformError

UpdatePartitionError

UpdateRegistryError

UpdateSchemaError

UpdateTableError

UpdateTriggerError

UpdateUserDefinedFunctionError

UpdateWorkflowError

ValidationException

A value could not be validated.

VersionMismatchException

There was a version conflict.

Enums

BatchCreatePartitionErrorKind

BatchDeleteConnectionErrorKind

BatchDeletePartitionErrorKind

BatchDeleteTableErrorKind

BatchDeleteTableVersionErrorKind

BatchGetBlueprintsErrorKind

BatchGetCrawlersErrorKind

BatchGetDevEndpointsErrorKind

BatchGetJobsErrorKind

BatchGetPartitionErrorKind

BatchGetTriggersErrorKind

BatchGetWorkflowsErrorKind

BatchStopJobRunErrorKind

BatchUpdatePartitionErrorKind

CancelMLTaskRunErrorKind

CheckSchemaVersionValidityErrorKind

CreateBlueprintErrorKind

CreateClassifierErrorKind

CreateConnectionErrorKind

CreateCrawlerErrorKind

CreateDatabaseErrorKind

CreateDevEndpointErrorKind

CreateJobErrorKind

CreateMLTransformErrorKind

CreatePartitionErrorKind

CreatePartitionIndexErrorKind

CreateRegistryErrorKind

CreateSchemaErrorKind

CreateScriptErrorKind

CreateSecurityConfigurationErrorKind

CreateTableErrorKind

CreateTriggerErrorKind

CreateUserDefinedFunctionErrorKind

CreateWorkflowErrorKind

DeleteBlueprintErrorKind

DeleteClassifierErrorKind

DeleteColumnStatisticsForPartitionErrorKind

DeleteColumnStatisticsForTableErrorKind

DeleteConnectionErrorKind

DeleteCrawlerErrorKind

DeleteDatabaseErrorKind

DeleteDevEndpointErrorKind

DeleteJobErrorKind

DeleteMLTransformErrorKind

DeletePartitionErrorKind

DeletePartitionIndexErrorKind

DeleteRegistryErrorKind

DeleteResourcePolicyErrorKind

DeleteSchemaErrorKind

DeleteSchemaVersionsErrorKind

DeleteSecurityConfigurationErrorKind

DeleteTableErrorKind

DeleteTableVersionErrorKind

DeleteTriggerErrorKind

DeleteUserDefinedFunctionErrorKind

DeleteWorkflowErrorKind

GetBlueprintErrorKind

GetBlueprintRunErrorKind

GetBlueprintRunsErrorKind

GetCatalogImportStatusErrorKind

GetClassifierErrorKind

GetClassifiersErrorKind

GetColumnStatisticsForPartitionErrorKind

GetColumnStatisticsForTableErrorKind

GetConnectionErrorKind

GetConnectionsErrorKind

GetCrawlerErrorKind

GetCrawlerMetricsErrorKind

GetCrawlersErrorKind

GetDataCatalogEncryptionSettingsErrorKind

GetDatabaseErrorKind

GetDatabasesErrorKind

GetDataflowGraphErrorKind

GetDevEndpointErrorKind

GetDevEndpointsErrorKind

GetJobBookmarkErrorKind

GetJobErrorKind

GetJobRunErrorKind

GetJobRunsErrorKind

GetJobsErrorKind

GetMLTaskRunErrorKind

GetMLTaskRunsErrorKind

GetMLTransformErrorKind

GetMLTransformsErrorKind

GetMappingErrorKind

GetPartitionErrorKind

GetPartitionIndexesErrorKind

GetPartitionsErrorKind

GetPlanErrorKind

GetRegistryErrorKind

GetResourcePoliciesErrorKind

GetResourcePolicyErrorKind

GetSchemaByDefinitionErrorKind

GetSchemaErrorKind

GetSchemaVersionErrorKind

GetSchemaVersionsDiffErrorKind

GetSecurityConfigurationErrorKind

GetSecurityConfigurationsErrorKind

GetTableErrorKind

GetTableVersionErrorKind

GetTableVersionsErrorKind

GetTablesErrorKind

GetTagsErrorKind

GetTriggerErrorKind

GetTriggersErrorKind

GetUserDefinedFunctionErrorKind

GetUserDefinedFunctionsErrorKind

GetWorkflowErrorKind

GetWorkflowRunErrorKind

GetWorkflowRunPropertiesErrorKind

GetWorkflowRunsErrorKind

ImportCatalogToGlueErrorKind

ListBlueprintsErrorKind

ListCrawlersErrorKind

ListDevEndpointsErrorKind

ListJobsErrorKind

ListMLTransformsErrorKind

ListRegistriesErrorKind

ListSchemaVersionsErrorKind

ListSchemasErrorKind

ListTriggersErrorKind

ListWorkflowsErrorKind

PutDataCatalogEncryptionSettingsErrorKind

PutResourcePolicyErrorKind

PutSchemaVersionMetadataErrorKind

PutWorkflowRunPropertiesErrorKind

QuerySchemaVersionMetadataErrorKind

RegisterSchemaVersionErrorKind

RemoveSchemaVersionMetadataErrorKind

ResetJobBookmarkErrorKind

ResumeWorkflowRunErrorKind

SearchTablesErrorKind

StartBlueprintRunErrorKind

StartCrawlerErrorKind

StartCrawlerScheduleErrorKind

StartExportLabelsTaskRunErrorKind

StartImportLabelsTaskRunErrorKind

StartJobRunErrorKind

StartMLEvaluationTaskRunErrorKind

StartMLLabelingSetGenerationTaskRunErrorKind

StartTriggerErrorKind

StartWorkflowRunErrorKind

StopCrawlerErrorKind

StopCrawlerScheduleErrorKind

StopTriggerErrorKind

StopWorkflowRunErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateBlueprintErrorKind

UpdateClassifierErrorKind

UpdateColumnStatisticsForPartitionErrorKind

UpdateColumnStatisticsForTableErrorKind

UpdateConnectionErrorKind

UpdateCrawlerErrorKind

UpdateCrawlerScheduleErrorKind

UpdateDatabaseErrorKind

UpdateDevEndpointErrorKind

UpdateJobErrorKind

UpdateMLTransformErrorKind

UpdatePartitionErrorKind

UpdateRegistryErrorKind

UpdateSchemaErrorKind

UpdateTableErrorKind

UpdateTriggerErrorKind

UpdateUserDefinedFunctionErrorKind

UpdateWorkflowErrorKind