Struct aws_sdk_glue::input::StartJobRunInput

#[non_exhaustive]
pub struct StartJobRunInput {
    pub job_name: Option<String>,
    pub job_run_id: Option<String>,
    pub arguments: Option<HashMap<String, String>>,
    pub allocated_capacity: i32,
    pub timeout: Option<i32>,
    pub max_capacity: Option<f64>,
    pub security_configuration: Option<String>,
    pub notification_property: Option<NotificationProperty>,
    pub worker_type: Option<WorkerType>,
    pub number_of_workers: Option<i32>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

job_name: Option<String>

The name of the job definition to use.

job_run_id: Option<String>

The ID of a previous JobRun to retry.

arguments: Option<HashMap<String, String>>

The job arguments specifically for this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

allocated_capacity: i32

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) to allocate to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

timeout: Option<i32>

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

max_capacity: Option<f64>

The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job, or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

security_configuration: Option<String>

The name of the SecurityConfiguration structure to be used with this job run.

notification_property: Option<NotificationProperty>

Specifies configuration properties of a job run notification.

worker_type: Option<WorkerType>

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

number_of_workers: Option<i32>

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

Implementations

impl StartJobRunInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<StartJobRun, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<StartJobRun>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture StartJobRunInput