Module aws_sdk_glue::model

Modules

action

See Action

backfill_error

See BackfillError

batch_stop_job_run_error

See BatchStopJobRunError

batch_stop_job_run_successful_submission

See BatchStopJobRunSuccessfulSubmission

batch_update_partition_failure_entry

See BatchUpdatePartitionFailureEntry

batch_update_partition_request_entry

See BatchUpdatePartitionRequestEntry

binary_column_statistics_data

See BinaryColumnStatisticsData

blueprint

See Blueprint

blueprint_details

See BlueprintDetails

blueprint_run

See BlueprintRun

boolean_column_statistics_data

See BooleanColumnStatisticsData

catalog_entry

See CatalogEntry

catalog_import_status

See CatalogImportStatus

catalog_target

See CatalogTarget

classifier

See Classifier

cloud_watch_encryption

See CloudWatchEncryption

code_gen_edge

See CodeGenEdge

code_gen_node

See CodeGenNode

code_gen_node_arg

See CodeGenNodeArg

column

See Column

column_error

See ColumnError

column_importance

See ColumnImportance

column_statistics

See ColumnStatistics

column_statistics_data

See ColumnStatisticsData

column_statistics_error

See ColumnStatisticsError

condition

See Condition

confusion_matrix

See ConfusionMatrix

connection

See Connection

connection_input

See ConnectionInput

connection_password_encryption

See ConnectionPasswordEncryption

connections_list

See ConnectionsList

crawl

See Crawl

crawler

See Crawler

crawler_metrics

See CrawlerMetrics

crawler_node_details

See CrawlerNodeDetails

crawler_targets

See CrawlerTargets

create_csv_classifier_request

See CreateCsvClassifierRequest

create_grok_classifier_request

See CreateGrokClassifierRequest

create_json_classifier_request

See CreateJsonClassifierRequest

create_xml_classifier_request

See CreateXmlClassifierRequest

csv_classifier

See CsvClassifier

data_catalog_encryption_settings

See DataCatalogEncryptionSettings

data_lake_principal

See DataLakePrincipal

database

See Database

database_identifier

See DatabaseIdentifier

database_input

See DatabaseInput

date_column_statistics_data

See DateColumnStatisticsData

decimal_column_statistics_data

See DecimalColumnStatisticsData

decimal_number

See DecimalNumber

dev_endpoint

See DevEndpoint

dev_endpoint_custom_libraries

See DevEndpointCustomLibraries

double_column_statistics_data

See DoubleColumnStatisticsData

dynamo_db_target

See DynamoDbTarget

edge

See Edge

encryption_at_rest

See EncryptionAtRest

encryption_configuration

See EncryptionConfiguration

error_detail

See ErrorDetail

error_details

See ErrorDetails

evaluation_metrics

See EvaluationMetrics

event_batching_condition

See EventBatchingCondition

execution_property

See ExecutionProperty

export_labels_task_run_properties

See ExportLabelsTaskRunProperties

find_matches_metrics

See FindMatchesMetrics

find_matches_parameters

See FindMatchesParameters

find_matches_task_run_properties

See FindMatchesTaskRunProperties

get_connections_filter

See GetConnectionsFilter

glue_policy

See GluePolicy

glue_table

See GlueTable

grok_classifier

See GrokClassifier

import_labels_task_run_properties

See ImportLabelsTaskRunProperties

jdbc_target

See JdbcTarget

job

See Job

job_bookmark_entry

See JobBookmarkEntry

job_bookmarks_encryption

See JobBookmarksEncryption

job_command

See JobCommand

job_node_details

See JobNodeDetails

job_run

See JobRun

job_update

See JobUpdate

json_classifier

See JsonClassifier

key_schema_element

See KeySchemaElement

labeling_set_generation_task_run_properties

See LabelingSetGenerationTaskRunProperties

last_active_definition

See LastActiveDefinition

last_crawl_info

See LastCrawlInfo

lineage_configuration

See LineageConfiguration

location

See Location

long_column_statistics_data

See LongColumnStatisticsData

mapping_entry

See MappingEntry

metadata_info

See MetadataInfo

metadata_key_value_pair

See MetadataKeyValuePair

ml_transform

See MlTransform

ml_user_data_encryption

See MlUserDataEncryption

mongo_db_target

See MongoDbTarget

node

See Node

notification_property

See NotificationProperty

order

See Order

other_metadata_value_list_item

See OtherMetadataValueListItem

partition

See Partition

partition_error

See PartitionError

partition_index

See PartitionIndex

partition_index_descriptor

See PartitionIndexDescriptor

partition_input

See PartitionInput

partition_value_list

See PartitionValueList

physical_connection_requirements

See PhysicalConnectionRequirements

predecessor

See Predecessor

predicate

See Predicate

principal_permissions

See PrincipalPermissions

property_predicate

See PropertyPredicate

recrawl_policy

See RecrawlPolicy

registry_id

See RegistryId

registry_list_item

See RegistryListItem

resource_uri

See ResourceUri

s3_encryption

See S3Encryption

s3_target

See S3Target

schedule

See Schedule

schema_change_policy

See SchemaChangePolicy

schema_column

See SchemaColumn

schema_id

See SchemaId

schema_list_item

See SchemaListItem

schema_reference

See SchemaReference

schema_version_error_item

See SchemaVersionErrorItem

schema_version_list_item

See SchemaVersionListItem

schema_version_number

See SchemaVersionNumber

security_configuration

See SecurityConfiguration

segment

See Segment

ser_de_info

See SerDeInfo

skewed_info

See SkewedInfo

sort_criterion

See SortCriterion

starting_event_batch_condition

See StartingEventBatchCondition

storage_descriptor

See StorageDescriptor

string_column_statistics_data

See StringColumnStatisticsData

table

See Table

table_error

See TableError

table_identifier

See TableIdentifier

table_input

See TableInput

table_version

See TableVersion

table_version_error

See TableVersionError

task_run

See TaskRun

task_run_filter_criteria

See TaskRunFilterCriteria

task_run_properties

See TaskRunProperties

task_run_sort_criteria

See TaskRunSortCriteria

transform_encryption

See TransformEncryption

transform_filter_criteria

See TransformFilterCriteria

transform_parameters

See TransformParameters

transform_sort_criteria

See TransformSortCriteria

trigger

See Trigger

trigger_node_details

See TriggerNodeDetails

trigger_update

See TriggerUpdate

update_csv_classifier_request

See UpdateCsvClassifierRequest

update_grok_classifier_request

See UpdateGrokClassifierRequest

update_json_classifier_request

See UpdateJsonClassifierRequest

update_xml_classifier_request

See UpdateXmlClassifierRequest

user_defined_function

See UserDefinedFunction

user_defined_function_input

See UserDefinedFunctionInput

workflow

See Workflow

workflow_graph

See WorkflowGraph

workflow_run

See WorkflowRun

workflow_run_statistics

See WorkflowRunStatistics

xml_classifier

See XmlClassifier

Structs

Action

Defines an action to be initiated by a trigger.

BackfillError

A list of errors that can occur when registering partition indexes for an existing table.

BatchStopJobRunError

Records an error that occurred when attempting to stop a

BatchStopJobRunSuccessfulSubmission

Records a successful request to stop a specified JobRun.

BatchUpdatePartitionFailureEntry

Contains information about a batch update partition error.

BatchUpdatePartitionRequestEntry

A structure that contains the values and structure used to update a partition.

BinaryColumnStatisticsData

Defines column statistics supported for bit sequence data values.

Blueprint

The details of a blueprint.

BlueprintDetails

The details of a blueprint.

BlueprintRun

The details of a blueprint run.

BooleanColumnStatisticsData

Defines column statistics supported for Boolean data columns.

CatalogEntry

Specifies a table definition in the Glue Data Catalog.

CatalogImportStatus

A structure containing migration status information.

CatalogTarget

Specifies an Glue Data Catalog target.

Classifier

Classifiers are triggered during a crawl task. A classifier checks whether a given file is

CloudWatchEncryption

Specifies how Amazon CloudWatch data should be encrypted.

CodeGenEdge

Represents a directional edge in a directed acyclic graph (DAG).

CodeGenNode

Represents a node in a directed acyclic graph (DAG)

CodeGenNodeArg

An argument or property of a node.

Column

A column in a Table.

ColumnError

Encapsulates a column name that failed and the reason for failure.

ColumnImportance

A structure containing the column name and column importance score for a column.

ColumnStatistics

Represents the generated column-level statistics for a table or partition.

ColumnStatisticsData

Contains the individual types of column statistics data. Only one data object should be set and indicated by the Type attribute.

ColumnStatisticsError

Encapsulates a ColumnStatistics object that failed and the reason for failure.

Condition

Defines a condition under which a trigger fires.

ConfusionMatrix

The confusion matrix shows you what your transform is predicting accurately and what types of errors it is making.

Connection

Defines a connection to a data source.

ConnectionInput

A structure that is used to specify a connection to create or update.

ConnectionPasswordEncryption

The data structure used by the Data Catalog to encrypt the password as part of

ConnectionsList

Specifies the connections used by a job.

Crawl

The details of a crawl in the workflow.

Crawler

Specifies a crawler program that examines a data source and uses classifiers to try to

CrawlerMetrics

Metrics for a specified crawler.

CrawlerNodeDetails

The details of a Crawler node present in the workflow.

CrawlerTargets

Specifies data stores to crawl.

CreateCsvClassifierRequest

Specifies a custom CSV classifier for CreateClassifier to create.

CreateGrokClassifierRequest

Specifies a grok classifier for CreateClassifier

CreateJsonClassifierRequest

Specifies a JSON classifier for CreateClassifier to create.

CreateXmlClassifierRequest

Specifies an XML classifier for CreateClassifier to create.

CsvClassifier

A classifier for custom CSV content.

DataCatalogEncryptionSettings

Contains configuration information for maintaining Data Catalog security.

DataLakePrincipal

The Lake Formation principal.

Database

The Database object represents a logical grouping of tables that might reside

DatabaseIdentifier

A structure that describes a target database for resource linking.

DatabaseInput

The structure used to create or update a database.

DateColumnStatisticsData

Defines column statistics supported for timestamp data columns.

DecimalColumnStatisticsData

Defines column statistics supported for fixed-point number data columns.

DecimalNumber

Contains a numeric value in decimal format.

DevEndpoint

A development endpoint where a developer can remotely debug extract, transform, and load

DevEndpointCustomLibraries

Custom libraries to be loaded into a development endpoint.

DoubleColumnStatisticsData

Defines column statistics supported for floating-point number data columns.

DynamoDbTarget

Specifies an Amazon DynamoDB table to crawl.

Edge

An edge represents a directed connection between two components

EncryptionAtRest

Specifies the encryption-at-rest configuration for the Data Catalog.

EncryptionConfiguration

Specifies an encryption configuration.

ErrorDetail

Contains details about an error.

ErrorDetails

An object containing error details.

EvaluationMetrics

Evaluation metrics provide an estimate of the quality of your machine learning transform.

EventBatchingCondition

Batch condition that must be met (specified number of events received or batch time window expired)

ExecutionProperty

An execution property of a job.

ExportLabelsTaskRunProperties

Specifies configuration properties for an exporting labels task run.

FindMatchesMetrics

The evaluation metrics for the find matches algorithm. The quality of your machine

FindMatchesParameters

The parameters to configure the find matches transform.

FindMatchesTaskRunProperties

Specifies configuration properties for a Find Matches task run.

GetConnectionsFilter

Filters the connection definitions that are returned by the GetConnections

GluePolicy

A structure for returning a resource policy.

GlueTable

The database and table in the Glue Data Catalog that is used for input or output data.

GrokClassifier

A classifier that uses grok patterns.

ImportLabelsTaskRunProperties

Specifies configuration properties for an importing labels task run.

JdbcTarget

Specifies a JDBC data store to crawl.

Job

Specifies a job definition.

JobBookmarkEntry

Defines a point that a job can resume processing.

JobBookmarksEncryption

Specifies how job bookmark data should be encrypted.

JobCommand

Specifies code that runs when a job is run.

JobNodeDetails

The details of a Job node present in the workflow.

JobRun

Contains information about a job run.

JobUpdate

Specifies information used to update an existing job definition. The previous job

JsonClassifier

A classifier for JSON content.

KeySchemaElement

A partition key pair consisting of a name and a type.

LabelingSetGenerationTaskRunProperties

Specifies configuration properties for a labeling set generation task run.

LastActiveDefinition

When there are multiple versions of a blueprint and the latest version has some errors, this attribute indicates the last successful blueprint definition that is available with the service.

LastCrawlInfo

Status and error information about the most recent crawl.

LineageConfiguration

Specifies data lineage configuration settings for the crawler.

Location

The location of resources.

LongColumnStatisticsData

Defines column statistics supported for integer data columns.

MappingEntry

Defines a mapping.

MetadataInfo

A structure containing metadata information for a schema version.

MetadataKeyValuePair

A structure containing a key value pair for metadata.

MlTransform

A structure for a machine learning transform.

MlUserDataEncryption

The encryption-at-rest settings of the transform that apply to accessing user data.

MongoDbTarget

Specifies an Amazon DocumentDB or MongoDB data store to crawl.

Node

A node represents an Glue component (trigger, crawler, or job) on a workflow graph.

NotificationProperty

Specifies configuration properties of a notification.

Order

Specifies the sort order of a sorted column.

OtherMetadataValueListItem

A structure containing other metadata for a schema version belonging to the same metadata key.

Partition

Represents a slice of table data.

PartitionError

Contains information about a partition error.

PartitionIndex

A structure for a partition index.

PartitionIndexDescriptor

A descriptor for a partition index in a table.

PartitionInput

The structure used to create and update a partition.

PartitionValueList

Contains a list of values defining partitions.

PhysicalConnectionRequirements

Specifies the physical requirements for a connection.

Predecessor

A job run that was used in the predicate of a conditional trigger

Predicate

Defines the predicate of the trigger, which determines when it fires.

PrincipalPermissions

Permissions granted to a principal.

PropertyPredicate

Defines a property predicate.

RecrawlPolicy

When crawling an Amazon S3 data source after the first crawl is complete, specifies whether to crawl the entire dataset again or to crawl only folders that were added since the last crawler run. For more information, see Incremental Crawls in Glue in the developer guide.

RegistryId

A wrapper structure that may contain the registry name and Amazon Resource Name (ARN).

RegistryListItem

A structure containing the details for a registry.

ResourceUri

The URIs for function resources.

S3Encryption

Specifies how Amazon Simple Storage Service (Amazon S3) data should be encrypted.

S3Target

Specifies a data store in Amazon Simple Storage Service (Amazon S3).

Schedule

A scheduling object using a cron statement to schedule an event.

SchemaChangePolicy

A policy that specifies update and deletion behaviors for the crawler.

SchemaColumn

A key-value pair representing a column and data type that this transform can

SchemaId

The unique ID of the schema in the Glue schema registry.

SchemaListItem

An object that contains minimal details for a schema.

SchemaReference

An object that references a schema stored in the Glue Schema Registry.

SchemaVersionErrorItem

An object that contains the error details for an operation on a schema version.

SchemaVersionListItem

An object containing the details about a schema version.

SchemaVersionNumber

A structure containing the schema version information.

SecurityConfiguration

Specifies a security configuration.

Segment

Defines a non-overlapping region of a table's partitions, allowing

SerDeInfo

Information about a serialization/deserialization program (SerDe) that serves as an

SkewedInfo

Specifies skewed values in a table. Skewed values are those that occur with very high

SortCriterion

Specifies a field to sort by and a sort order.

StartingEventBatchCondition

The batch condition that started the workflow run. Either the number of events in the batch size arrived,

StorageDescriptor

Describes the physical storage of table data.

StringColumnStatisticsData

Defines column statistics supported for character sequence data values.

Table

Represents a collection of related data organized in columns and rows.

TableError

An error record for table operations.

TableIdentifier

A structure that describes a target table for resource linking.

TableInput

A structure used to define a table.

TableVersion

Specifies a version of a table.

TableVersionError

An error record for table-version operations.

TaskRun

The sampling parameters that are associated with the machine learning transform.

TaskRunFilterCriteria

The criteria that are used to filter the task runs for the machine learning

TaskRunProperties

The configuration properties for the task run.

TaskRunSortCriteria

The sorting criteria that are used to sort the list of task runs for the machine learning

TransformEncryption

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

TransformFilterCriteria

The criteria used to filter the machine learning transforms.

TransformParameters

The algorithm-specific parameters that are associated with the machine learning

TransformSortCriteria

The sorting criteria that are associated with the machine learning transform.

Trigger

Information about a specific trigger.

TriggerNodeDetails

The details of a Trigger node present in the workflow.

TriggerUpdate

A structure used to provide information used to update a trigger. This object updates the

UpdateCsvClassifierRequest

Specifies a custom CSV classifier to be updated.

UpdateGrokClassifierRequest

Specifies a grok classifier to update when passed to

UpdateJsonClassifierRequest

Specifies a JSON classifier to be updated.

UpdateXmlClassifierRequest

Specifies an XML classifier to be updated.

UserDefinedFunction

Represents the equivalent of a Hive user-defined function

UserDefinedFunctionInput

A structure used to create or update a user-defined function.

Workflow

A workflow is a collection of multiple dependent Glue

WorkflowGraph

A workflow graph represents the complete workflow containing all the Glue components present in the

WorkflowRun

A workflow run is an execution of a workflow providing all the runtime information.

WorkflowRunStatistics

Workflow run statistics provides statistics about the workflow run.

XmlClassifier

A classifier for XML content.

Enums

BackfillErrorCode

BlueprintRunState

BlueprintStatus

CatalogEncryptionMode

CloudWatchEncryptionMode

ColumnStatisticsType

Comparator

Compatibility

ConnectionPropertyKey

ConnectionType

CrawlState

CrawlerLineageSettings

CrawlerState

CsvHeaderOption

NOTE: CsvHeaderOption::Unknown has been renamed to ::UnknownValue.

DataFormat

DeleteBehavior

EnableHybridValues

ExistCondition

JobBookmarksEncryptionMode

JobRunState

Language

LastCrawlStatus

Logical

LogicalOperator

MlUserDataEncryptionModeString

NodeType

PartitionIndexStatus

Permission

PrincipalType

RecrawlBehavior

RegistryStatus

ResourceShareType

ResourceType

S3EncryptionMode

ScheduleState

SchemaDiffType

SchemaStatus

SchemaVersionStatus

Sort

SortDirectionType

TaskRunSortColumnType

TaskStatusType

TaskType

TransformSortColumnType

TransformStatusType

TransformType

TriggerState

TriggerType

UpdateBehavior

WorkerType

WorkflowRunStatus