Struct aws_sdk_glue::model::crawler::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Crawler

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the crawler.

pub fn set_name(self, input: Option<String>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an IAM role that's used to access customer resources, such as Amazon Simple Storage Service (Amazon S3) data.

pub fn set_role(self, input: Option<String>) -> Self

pub fn targets(self, input: CrawlerTargets) -> Self

A collection of targets to crawl.

pub fn set_targets(self, input: Option<CrawlerTargets>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The name of the database in which the crawler's output is stored.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description of the crawler.

pub fn set_description(self, input: Option<String>) -> Self

pub fn classifiers(self, input: impl Into<String>) -> Self

pub fn set_classifiers(self, input: Option<Vec<String>>) -> Self

pub fn recrawl_policy(self, input: RecrawlPolicy) -> Self

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

pub fn set_recrawl_policy(self, input: Option<RecrawlPolicy>) -> Self

pub fn schema_change_policy(self, input: SchemaChangePolicy) -> Self

The policy that specifies update and delete behaviors for the crawler.

pub fn set_schema_change_policy(self, input: Option<SchemaChangePolicy>) -> Self

pub fn lineage_configuration(self, input: LineageConfiguration) -> Self

A configuration that specifies whether data lineage is enabled for the crawler.

pub fn set_lineage_configuration(
self,
input: Option<LineageConfiguration>
) -> Self

pub fn state(self, input: CrawlerState) -> Self

Indicates whether the crawler is running, or whether a run is pending.

pub fn set_state(self, input: Option<CrawlerState>) -> Self

pub fn table_prefix(self, input: impl Into<String>) -> Self

The prefix added to the names of tables that are created.

pub fn set_table_prefix(self, input: Option<String>) -> Self

pub fn schedule(self, input: Schedule) -> Self

For scheduled crawlers, the schedule when the crawler runs.

pub fn set_schedule(self, input: Option<Schedule>) -> Self

pub fn crawl_elapsed_time(self, input: i64) -> Self

If the crawler is running, contains the total time elapsed since the last crawl began.

pub fn set_crawl_elapsed_time(self, input: Option<i64>) -> Self

pub fn creation_time(self, input: Instant) -> Self

The time that the crawler was created.

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn last_updated(self, input: Instant) -> Self

The time that the crawler was last updated.

pub fn set_last_updated(self, input: Option<Instant>) -> Self

pub fn last_crawl(self, input: LastCrawlInfo) -> Self

The status of the last crawl, and potentially error information if an error occurred.

pub fn set_last_crawl(self, input: Option<LastCrawlInfo>) -> Self

pub fn version(self, input: i64) -> Self

The version of the crawler.

pub fn set_version(self, input: Option<i64>) -> Self

pub fn configuration(self, input: impl Into<String>) -> Self

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Include and Exclude Patterns.

pub fn set_configuration(self, input: Option<String>) -> Self

pub fn crawler_security_configuration(self, input: impl Into<String>) -> Self

The name of the SecurityConfiguration structure to be used by this crawler.

pub fn set_crawler_security_configuration(self, input: Option<String>) -> Self

pub fn build(self) -> Crawler

Consumes the builder and constructs a Crawler