Struct aws_sdk_glue::model::dev_endpoint::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DevEndpoint

Implementations

impl Builder

pub fn endpoint_name(self, input: impl Into<String>) -> Self

The name of the DevEndpoint.

pub fn set_endpoint_name(self, input: Option<String>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM role used in this DevEndpoint.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn subnet_id(self, input: impl Into<String>) -> Self

The subnet ID for this DevEndpoint.

pub fn set_subnet_id(self, input: Option<String>) -> Self

pub fn yarn_endpoint_address(self, input: impl Into<String>) -> Self

The YARN endpoint address used by this DevEndpoint.

pub fn set_yarn_endpoint_address(self, input: Option<String>) -> Self

pub fn private_address(self, input: impl Into<String>) -> Self

A private IP address to access the DevEndpoint within a VPC if the DevEndpoint is created within one. The PrivateAddress field is present only when you create the DevEndpoint within your VPC.

pub fn set_private_address(self, input: Option<String>) -> Self

pub fn zeppelin_remote_spark_interpreter_port(self, input: i32) -> Self

The Apache Zeppelin port for the remote Apache Spark interpreter.

pub fn set_zeppelin_remote_spark_interpreter_port(
self,
input: Option<i32>
) -> Self

pub fn public_address(self, input: impl Into<String>) -> Self

The public IP address used by this DevEndpoint. The PublicAddress field is present only when you create a non-virtual private cloud (VPC) DevEndpoint.

pub fn set_public_address(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The current status of this DevEndpoint.

pub fn set_status(self, input: Option<String>) -> Self

pub fn worker_type(self, input: WorkerType) -> Self

The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

pub fn set_worker_type(self, input: Option<WorkerType>) -> Self

pub fn glue_version(self, input: impl Into<String>) -> Self

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

pub fn set_glue_version(self, input: Option<String>) -> Self

pub fn number_of_workers(self, input: i32) -> Self

The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

pub fn set_number_of_workers(self, input: Option<i32>) -> Self

pub fn number_of_nodes(self, input: i32) -> Self

The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

pub fn set_number_of_nodes(self, input: Option<i32>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The AWS Availability Zone where this DevEndpoint is located.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The ID of the virtual private cloud (VPC) used by this DevEndpoint.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn extra_python_libs_s3_path(self, input: impl Into<String>) -> Self

The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not currently supported.

pub fn set_extra_python_libs_s3_path(self, input: Option<String>) -> Self

pub fn extra_jars_s3_path(self, input: impl Into<String>) -> Self

The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

You can only use pure Java/Scala libraries with a DevEndpoint.

pub fn set_extra_jars_s3_path(self, input: Option<String>) -> Self

pub fn failure_reason(self, input: impl Into<String>) -> Self

The reason for a current failure in this DevEndpoint.

pub fn set_failure_reason(self, input: Option<String>) -> Self

pub fn last_update_status(self, input: impl Into<String>) -> Self

The status of the last update.

pub fn set_last_update_status(self, input: Option<String>) -> Self

pub fn created_timestamp(self, input: Instant) -> Self

The point in time at which this DevEndpoint was created.

pub fn set_created_timestamp(self, input: Option<Instant>) -> Self

pub fn last_modified_timestamp(self, input: Instant) -> Self

The point in time at which this DevEndpoint was last modified.

pub fn set_last_modified_timestamp(self, input: Option<Instant>) -> Self

pub fn public_key(self, input: impl Into<String>) -> Self

The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

pub fn set_public_key(self, input: Option<String>) -> Self

pub fn public_keys(self, input: impl Into<String>) -> Self

pub fn set_public_keys(self, input: Option<Vec<String>>) -> Self

pub fn security_configuration(self, input: impl Into<String>) -> Self

The name of the SecurityConfiguration structure to be used with this DevEndpoint.

pub fn set_security_configuration(self, input: Option<String>) -> Self

pub fn arguments(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_arguments(self, input: Option<HashMap<String, String>>) -> Self

pub fn build(self) -> DevEndpoint

Consumes the builder and constructs a DevEndpoint