Struct aws_sdk_glue::model::job::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Job

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name you assign to this job definition.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description of the job.

pub fn set_description(self, input: Option<String>) -> Self

pub fn log_uri(self, input: impl Into<String>) -> Self

This field is reserved for future use.

pub fn set_log_uri(self, input: Option<String>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

pub fn set_role(self, input: Option<String>) -> Self

pub fn created_on(self, input: Instant) -> Self

The time and date that this job definition was created.

pub fn set_created_on(self, input: Option<Instant>) -> Self

pub fn last_modified_on(self, input: Instant) -> Self

The last point in time when this job definition was modified.

pub fn set_last_modified_on(self, input: Option<Instant>) -> Self

pub fn execution_property(self, input: ExecutionProperty) -> Self

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

pub fn set_execution_property(self, input: Option<ExecutionProperty>) -> Self

pub fn command(self, input: JobCommand) -> Self

The JobCommand that runs this job.

pub fn set_command(self, input: Option<JobCommand>) -> Self

pub fn default_arguments(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_default_arguments(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn non_overridable_arguments(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_non_overridable_arguments(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn connections(self, input: ConnectionsList) -> Self

The connections used for this job.

pub fn set_connections(self, input: Option<ConnectionsList>) -> Self

pub fn max_retries(self, input: i32) -> Self

The maximum number of times to retry this job after a JobRun fails.

pub fn set_max_retries(self, input: Option<i32>) -> Self

pub fn allocated_capacity(self, input: i32) -> Self

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to runs of this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

pub fn set_allocated_capacity(self, input: Option<i32>) -> Self

pub fn timeout(self, input: i32) -> Self

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

pub fn set_timeout(self, input: Option<i32>) -> Self

pub fn max_capacity(self, input: f64) -> Self

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job, an Apache Spark ETL job, or an Apache Spark streaming ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

pub fn set_max_capacity(self, input: Option<f64>) -> Self

pub fn worker_type(self, input: WorkerType) -> Self

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

pub fn set_worker_type(self, input: Option<WorkerType>) -> Self

pub fn number_of_workers(self, input: i32) -> Self

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

pub fn set_number_of_workers(self, input: Option<i32>) -> Self

pub fn security_configuration(self, input: impl Into<String>) -> Self

The name of the SecurityConfiguration structure to be used with this job.

pub fn set_security_configuration(self, input: Option<String>) -> Self

pub fn notification_property(self, input: NotificationProperty) -> Self

Specifies configuration properties of a job notification.

pub fn set_notification_property(
self,
input: Option<NotificationProperty>
) -> Self

pub fn glue_version(self, input: impl Into<String>) -> Self

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

pub fn set_glue_version(self, input: Option<String>) -> Self

pub fn build(self) -> Job

Consumes the builder and constructs a Job