Struct aws_sdk_glue::model::job_run::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for JobRun

Implementations

impl Builder

pub fn id(self, input: impl Into<String>) -> Self

The ID of this job run.

pub fn set_id(self, input: Option<String>) -> Self

The ID of this job run.

pub fn attempt(self, input: i32) -> Self

The number of the attempt to run this job.

pub fn set_attempt(self, input: Option<i32>) -> Self

The number of the attempt to run this job.

pub fn previous_run_id(self, input: impl Into<String>) -> Self

The ID of the previous run of this job. For example, the JobRunId specified in the StartJobRun action.

pub fn set_previous_run_id(self, input: Option<String>) -> Self

The ID of the previous run of this job. For example, the JobRunId specified in the StartJobRun action.

pub fn trigger_name(self, input: impl Into<String>) -> Self

The name of the trigger that started this job run.

pub fn set_trigger_name(self, input: Option<String>) -> Self

The name of the trigger that started this job run.

pub fn job_name(self, input: impl Into<String>) -> Self

The name of the job definition being used in this run.

pub fn set_job_name(self, input: Option<String>) -> Self

The name of the job definition being used in this run.

pub fn started_on(self, input: Instant) -> Self

The date and time at which this job run was started.

pub fn set_started_on(self, input: Option<Instant>) -> Self

The date and time at which this job run was started.

pub fn last_modified_on(self, input: Instant) -> Self

The last time that this job run was modified.

pub fn set_last_modified_on(self, input: Option<Instant>) -> Self

The last time that this job run was modified.

pub fn completed_on(self, input: Instant) -> Self

The date and time that this job run completed.

pub fn set_completed_on(self, input: Option<Instant>) -> Self

The date and time that this job run completed.

pub fn job_run_state(self, input: JobRunState) -> Self

The current state of the job run. For more information about the statuses of jobs that have terminated abnormally, see Glue Job Run Statuses.

pub fn set_job_run_state(self, input: Option<JobRunState>) -> Self

The current state of the job run. For more information about the statuses of jobs that have terminated abnormally, see Glue Job Run Statuses.

pub fn arguments(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to arguments.

To override the contents of this collection use set_arguments.

The job arguments associated with this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

pub fn set_arguments(self, input: Option<HashMap<String, String>>) -> Self

The job arguments associated with this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

pub fn error_message(self, input: impl Into<String>) -> Self

An error message associated with this job run.

pub fn set_error_message(self, input: Option<String>) -> Self

An error message associated with this job run.

pub fn predecessor_runs(self, input: impl Into<Predecessor>) -> Self

Appends an item to predecessor_runs.

To override the contents of this collection use set_predecessor_runs.

A list of predecessors to this job run.

pub fn set_predecessor_runs(self, input: Option<Vec<Predecessor>>) -> Self

A list of predecessors to this job run.

pub fn allocated_capacity(self, input: i32) -> Self

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

pub fn set_allocated_capacity(self, input: Option<i32>) -> Self

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

pub fn execution_time(self, input: i32) -> Self

The amount of time (in seconds) that the job run consumed resources.

pub fn set_execution_time(self, input: Option<i32>) -> Self

The amount of time (in seconds) that the job run consumed resources.

pub fn timeout(self, input: i32) -> Self

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

pub fn set_timeout(self, input: Option<i32>) -> Self

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

pub fn max_capacity(self, input: f64) -> Self

The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

pub fn set_max_capacity(self, input: Option<f64>) -> Self

The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

pub fn worker_type(self, input: WorkerType) -> Self

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

pub fn set_worker_type(self, input: Option<WorkerType>) -> Self

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

pub fn number_of_workers(self, input: i32) -> Self

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

pub fn set_number_of_workers(self, input: Option<i32>) -> Self

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

pub fn security_configuration(self, input: impl Into<String>) -> Self

The name of the SecurityConfiguration structure to be used with this job run.

pub fn set_security_configuration(self, input: Option<String>) -> Self

The name of the SecurityConfiguration structure to be used with this job run.

pub fn log_group_name(self, input: impl Into<String>) -> Self

The name of the log group for secure logging that can be server-side encrypted in Amazon CloudWatch using KMS. This name can be /aws-glue/jobs/, in which case the default encryption is NONE. If you add a role name and SecurityConfiguration name (in other words, /aws-glue/jobs-yourRoleName-yourSecurityConfigurationName/), then that security configuration is used to encrypt the log group.

pub fn set_log_group_name(self, input: Option<String>) -> Self

The name of the log group for secure logging that can be server-side encrypted in Amazon CloudWatch using KMS. This name can be /aws-glue/jobs/, in which case the default encryption is NONE. If you add a role name and SecurityConfiguration name (in other words, /aws-glue/jobs-yourRoleName-yourSecurityConfigurationName/), then that security configuration is used to encrypt the log group.

pub fn notification_property(self, input: NotificationProperty) -> Self

Specifies configuration properties of a job run notification.

pub fn set_notification_property(
self,
input: Option<NotificationProperty>
) -> Self

Specifies configuration properties of a job run notification.

pub fn glue_version(self, input: impl Into<String>) -> Self

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

pub fn set_glue_version(self, input: Option<String>) -> Self

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

pub fn build(self) -> JobRun

Consumes the builder and constructs a JobRun