Struct aws_sdk_glue::model::ml_transform::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MlTransform

Implementations

impl Builder

pub fn transform_id(self, input: impl Into<String>) -> Self

The unique transform ID that is generated for the machine learning transform. The ID is guaranteed to be unique and does not change.

pub fn set_transform_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A user-defined name for the machine learning transform. Names are not guaranteed unique and can be changed at any time.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A user-defined, long-form description text for the machine learning transform. Descriptions are not guaranteed to be unique and can be changed at any time.

pub fn set_description(self, input: Option<String>) -> Self

pub fn status(self, input: TransformStatusType) -> Self

The current status of the machine learning transform.

pub fn set_status(self, input: Option<TransformStatusType>) -> Self

pub fn created_on(self, input: Instant) -> Self

A timestamp. The time and date that this machine learning transform was created.

pub fn set_created_on(self, input: Option<Instant>) -> Self

pub fn last_modified_on(self, input: Instant) -> Self

A timestamp. The last point in time when this machine learning transform was modified.

pub fn set_last_modified_on(self, input: Option<Instant>) -> Self

pub fn input_record_tables(self, input: impl Into<GlueTable>) -> Self

pub fn set_input_record_tables(self, input: Option<Vec<GlueTable>>) -> Self

pub fn parameters(self, input: TransformParameters) -> Self

A TransformParameters object. You can use parameters to tune (customize) the behavior of the machine learning transform by specifying what data it learns from and your preference on various tradeoffs (such as precious vs. recall, or accuracy vs. cost).

pub fn set_parameters(self, input: Option<TransformParameters>) -> Self

pub fn evaluation_metrics(self, input: EvaluationMetrics) -> Self

An EvaluationMetrics object. Evaluation metrics provide an estimate of the quality of your machine learning transform.

pub fn set_evaluation_metrics(self, input: Option<EvaluationMetrics>) -> Self

pub fn label_count(self, input: i32) -> Self

A count identifier for the labeling files generated by Glue for this transform. As you create a better transform, you can iteratively download, label, and upload the labeling file.

pub fn set_label_count(self, input: Option<i32>) -> Self

pub fn schema(self, input: impl Into<SchemaColumn>) -> Self

pub fn set_schema(self, input: Option<Vec<SchemaColumn>>) -> Self

pub fn role(self, input: impl Into<String>) -> Self

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions. The required permissions include both Glue service role permissions to Glue resources, and Amazon S3 permissions required by the transform.

This role needs Glue service role permissions to allow access to resources in Glue. See Attach a Policy to IAM Users That Access Glue.

This role needs permission to your Amazon Simple Storage Service (Amazon S3) sources, targets, temporary directory, scripts, and any libraries used by the task run for this transform.

pub fn set_role(self, input: Option<String>) -> Self

pub fn glue_version(self, input: impl Into<String>) -> Self

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

pub fn set_glue_version(self, input: Option<String>) -> Self

pub fn max_capacity(self, input: f64) -> Self

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

pub fn set_max_capacity(self, input: Option<f64>) -> Self

pub fn worker_type(self, input: WorkerType) -> Self

The type of predefined worker that is allocated when a task of this transform runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

pub fn set_worker_type(self, input: Option<WorkerType>) -> Self

pub fn number_of_workers(self, input: i32) -> Self

The number of workers of a defined workerType that are allocated when a task of the transform runs.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

pub fn set_number_of_workers(self, input: Option<i32>) -> Self

pub fn timeout(self, input: i32) -> Self

The timeout in minutes of the machine learning transform.

pub fn set_timeout(self, input: Option<i32>) -> Self

pub fn max_retries(self, input: i32) -> Self

The maximum number of times to retry after an MLTaskRun of the machine learning transform fails.

pub fn set_max_retries(self, input: Option<i32>) -> Self

pub fn transform_encryption(self, input: TransformEncryption) -> Self

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

pub fn set_transform_encryption(
self,
input: Option<TransformEncryption>
) -> Self

pub fn build(self) -> MlTransform

Consumes the builder and constructs a MlTransform