Struct aws_sdk_glue::model::workflow_run::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for WorkflowRun

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

Name of the workflow that was run.

pub fn set_name(self, input: Option<String>) -> Self

pub fn workflow_run_id(self, input: impl Into<String>) -> Self

The ID of this workflow run.

pub fn set_workflow_run_id(self, input: Option<String>) -> Self

pub fn previous_run_id(self, input: impl Into<String>) -> Self

The ID of the previous workflow run.

pub fn set_previous_run_id(self, input: Option<String>) -> Self

pub fn workflow_run_properties(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_workflow_run_properties(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn started_on(self, input: Instant) -> Self

The date and time when the workflow run was started.

pub fn set_started_on(self, input: Option<Instant>) -> Self

pub fn completed_on(self, input: Instant) -> Self

The date and time when the workflow run completed.

pub fn set_completed_on(self, input: Option<Instant>) -> Self

pub fn status(self, input: WorkflowRunStatus) -> Self

The status of the workflow run.

pub fn set_status(self, input: Option<WorkflowRunStatus>) -> Self

pub fn error_message(self, input: impl Into<String>) -> Self

This error message describes any error that may have occurred in starting the workflow run. Currently the only error message is "Concurrent runs exceeded for workflow: foo."

pub fn set_error_message(self, input: Option<String>) -> Self

pub fn statistics(self, input: WorkflowRunStatistics) -> Self

The statistics of the run.

pub fn set_statistics(self, input: Option<WorkflowRunStatistics>) -> Self

pub fn graph(self, input: WorkflowGraph) -> Self

The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

pub fn set_graph(self, input: Option<WorkflowGraph>) -> Self

pub fn starting_event_batch_condition(
self,
input: StartingEventBatchCondition
) -> Self

The batch condition that started the workflow run.

pub fn set_starting_event_batch_condition(
self,
input: Option<StartingEventBatchCondition>
) -> Self

pub fn build(self) -> WorkflowRun

Consumes the builder and constructs a WorkflowRun