Module aws_sdk_glue::operation

Structs

BatchCreatePartition

Creates one or more partitions in a batch operation.

BatchDeleteConnection

Deletes a list of connection definitions from the Data Catalog.

BatchDeletePartition

Deletes one or more partitions in a batch operation.

BatchDeleteTable

Deletes multiple tables at once.

BatchDeleteTableVersion

Deletes a specified batch of versions of a table.

BatchGetBlueprints

Retrieves information about a list of blueprints.

BatchGetCrawlers

Returns a list of resource metadata for a given list of crawler names. After calling the ListCrawlers operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

BatchGetDevEndpoints

Returns a list of resource metadata for a given list of development endpoint names. After

BatchGetJobs

Returns a list of resource metadata for a given list of job names. After calling the ListJobs operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

BatchGetPartition

Retrieves partitions in a batch request.

BatchGetTriggers

Returns a list of resource metadata for a given list of trigger names. After calling the ListTriggers operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

BatchGetWorkflows

Returns a list of resource metadata for a given list of workflow names. After calling the ListWorkflows operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

BatchStopJobRun

Stops one or more job runs for a specified job definition.

BatchUpdatePartition

Updates one or more partitions in a batch operation.

CancelMLTaskRun

Cancels (stops) a task run. Machine learning task runs are asynchronous tasks that Glue runs on your behalf as part of various machine learning workflows. You can cancel a

CheckSchemaVersionValidity

Validates the supplied schema. This call has no side effects, it simply validates using the supplied schema using DataFormat as the format. Since it does not take a schema set name, no compatibility checks are performed.

CreateBlueprint

Registers a blueprint with Glue.

CreateClassifier

Creates a classifier in the user's account. This can be a GrokClassifier, an

CreateConnection

Creates a connection definition in the Data Catalog.

CreateCrawler

Creates a new crawler with specified targets, role, configuration, and optional schedule.

CreateDatabase

Creates a new database in a Data Catalog.

CreateDevEndpoint

Creates a new development endpoint.

CreateJob

Creates a new job definition.

CreateMLTransform

Creates an Glue machine learning transform. This operation creates the transform and

CreatePartition

Creates a new partition.

CreatePartitionIndex

Creates a specified partition index in an existing table.

CreateRegistry

Creates a new registry which may be used to hold a collection of schemas.

CreateSchema

Creates a new schema set and registers the schema definition. Returns an error if the schema set already exists without actually registering the version.

CreateScript

Transforms a directed acyclic graph (DAG) into code.

CreateSecurityConfiguration

Creates a new security configuration. A security configuration is a set of security properties that can be used by Glue. You can use a security configuration to encrypt data at rest. For information about using security configurations in Glue, see Encrypting Data Written by Crawlers, Jobs, and Development Endpoints.

CreateTable

Creates a new table definition in the Data Catalog.

CreateTrigger

Creates a new trigger.

CreateUserDefinedFunction

Creates a new function definition in the Data Catalog.

CreateWorkflow

Creates a new workflow.

DeleteBlueprint

Deletes an existing blueprint.

DeleteClassifier

Removes a classifier from the Data Catalog.

DeleteColumnStatisticsForPartition

Delete the partition column statistics of a column.

DeleteColumnStatisticsForTable

Retrieves table statistics of columns.

DeleteConnection

Deletes a connection from the Data Catalog.

DeleteCrawler

Removes a specified crawler from the Glue Data Catalog, unless the crawler state is

DeleteDatabase

Removes a specified database from a Data Catalog.

DeleteDevEndpoint

Deletes a specified development endpoint.

DeleteJob

Deletes a specified job definition. If the job definition

DeleteMLTransform

Deletes an Glue machine learning transform. Machine learning transforms are a special

DeletePartition

Deletes a specified partition.

DeletePartitionIndex

Deletes a specified partition index from an existing table.

DeleteRegistry

Delete the entire registry including schema and all of its versions. To get the status of the delete operation, you can call the GetRegistry API after the asynchronous call. Deleting a registry will deactivate all online operations for the registry such as the UpdateRegistry, CreateSchema, UpdateSchema, and RegisterSchemaVersion APIs.

DeleteResourcePolicy

Deletes a specified policy.

DeleteSchema

Deletes the entire schema set, including the schema set and all of its versions. To get the status of the delete operation, you can call GetSchema API after the asynchronous call. Deleting a registry will deactivate all online operations for the schema, such as the GetSchemaByDefinition, and RegisterSchemaVersion APIs.

DeleteSchemaVersions

Remove versions from the specified schema. A version number or range may be supplied. If the compatibility mode forbids deleting of a version that is necessary, such as BACKWARDS_FULL, an error is returned. Calling the GetSchemaVersions API after this call will list the status of the deleted versions.

DeleteSecurityConfiguration

Deletes a specified security configuration.

DeleteTable

Removes a table definition from the Data Catalog.

DeleteTableVersion

Deletes a specified version of a table.

DeleteTrigger

Deletes a specified trigger. If the trigger is not found, no

DeleteUserDefinedFunction

Deletes an existing function definition from the Data Catalog.

DeleteWorkflow

Deletes a workflow.

GetBlueprint

Retrieves the details of a blueprint.

GetBlueprintRun

Retrieves the details of a blueprint run.

GetBlueprintRuns

Retrieves the details of blueprint runs for a specified blueprint.

GetCatalogImportStatus

Retrieves the status of a migration operation.

GetClassifier

Retrieve a classifier by name.

GetClassifiers

Lists all classifier objects in the Data Catalog.

GetColumnStatisticsForPartition

Retrieves partition statistics of columns.

GetColumnStatisticsForTable

Retrieves table statistics of columns.

GetConnection

Retrieves a connection definition from the Data Catalog.

GetConnections

Retrieves a list of connection definitions from the Data Catalog.

GetCrawler

Retrieves metadata for a specified crawler.

GetCrawlerMetrics

Retrieves metrics about specified crawlers.

GetCrawlers

Retrieves metadata for all crawlers defined in the customer

GetDataCatalogEncryptionSettings

Retrieves the security configuration for a specified catalog.

GetDatabase

Retrieves the definition of a specified database.

GetDatabases

Retrieves all databases defined in a given Data Catalog.

GetDataflowGraph

Transforms a Python script into a directed acyclic graph (DAG).

GetDevEndpoint

Retrieves information about a specified development endpoint.

GetDevEndpoints

Retrieves all the development endpoints in this AWS account.

GetJob

Retrieves an existing job definition.

GetJobBookmark

Returns information on a job bookmark entry.

GetJobRun

Retrieves the metadata for a given job run.

GetJobRuns

Retrieves metadata for all runs of a given job definition.

GetJobs

Retrieves all current job definitions.

GetMLTaskRun

Gets details for a specific task run on a machine learning transform. Machine learning

GetMLTaskRuns

Gets a list of runs for a machine learning transform. Machine learning task runs are

GetMLTransform

Gets an Glue machine learning transform artifact and all its corresponding metadata.

GetMLTransforms

Gets a sortable, filterable list of existing Glue machine learning transforms. Machine

GetMapping

Creates mappings.

GetPartition

Retrieves information about a specified partition.

GetPartitionIndexes

Retrieves the partition indexes associated with a table.

GetPartitions

Retrieves information about the partitions in a table.

GetPlan

Gets code to perform a specified mapping.

GetRegistry

Describes the specified registry in detail.

GetResourcePolicies

Retrieves the resource policies set on individual resources by Resource Access Manager

GetResourcePolicy

Retrieves a specified resource policy.

GetSchema

Describes the specified schema in detail.

GetSchemaByDefinition

Retrieves a schema by the SchemaDefinition. The schema definition is sent to the Schema Registry, canonicalized, and hashed. If the hash is matched within the scope of the SchemaName or ARN (or the default registry, if none is supplied), that schema’s metadata is returned. Otherwise, a 404 or NotFound error is returned. Schema versions in Deleted statuses will not be included in the results.

GetSchemaVersion

Get the specified schema by its unique ID assigned when a version of the schema is created or registered. Schema versions in Deleted status will not be included in the results.

GetSchemaVersionsDiff

Fetches the schema version difference in the specified difference type between two stored schema versions in the Schema Registry.

GetSecurityConfiguration

Retrieves a specified security configuration.

GetSecurityConfigurations

Retrieves a list of all security configurations.

GetTable

Retrieves the Table definition in a Data Catalog for

GetTableVersion

Retrieves a specified version of a table.

GetTableVersions

Retrieves a list of strings that identify available versions of

GetTables

Retrieves the definitions of some or all of the tables in a given

GetTags

Retrieves a list of tags associated with a resource.

GetTrigger

Retrieves the definition of a trigger.

GetTriggers

Gets all the triggers associated with a job.

GetUserDefinedFunction

Retrieves a specified function definition from the Data Catalog.

GetUserDefinedFunctions

Retrieves multiple function definitions from the Data Catalog.

GetWorkflow

Retrieves resource metadata for a workflow.

GetWorkflowRun

Retrieves the metadata for a given workflow run.

GetWorkflowRunProperties

Retrieves the workflow run properties which were set during the run.

GetWorkflowRuns

Retrieves metadata for all runs of a given workflow.

ImportCatalogToGlue

Imports an existing Amazon Athena Data Catalog to Glue.

ListBlueprints

Lists all the blueprint names in an account.

ListCrawlers

Retrieves the names of all crawler resources in this Amazon Web Services account, or the

ListDevEndpoints

Retrieves the names of all DevEndpoint resources in this Amazon Web Services account, or the

ListJobs

Retrieves the names of all job resources in this Amazon Web Services account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

ListMLTransforms

Retrieves a sortable, filterable list of existing Glue machine learning transforms in this Amazon Web Services account,

ListRegistries

Returns a list of registries that you have created, with minimal registry information. Registries in the Deleting status will not be included in the results. Empty results will be returned if there are no registries available.

ListSchemaVersions

Returns a list of schema versions that you have created, with minimal information. Schema versions in Deleted status will not be included in the results. Empty results will be returned if there are no schema versions available.

ListSchemas

Returns a list of schemas with minimal details. Schemas in Deleting status will not be included in the results. Empty results will be returned if there are no schemas available.

ListTriggers

Retrieves the names of all trigger resources in this Amazon Web Services account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

ListWorkflows

Lists names of workflows created in the account.

PutDataCatalogEncryptionSettings

Sets the security configuration for a specified catalog. After the configuration has been

PutResourcePolicy

Sets the Data Catalog resource policy for access control.

PutSchemaVersionMetadata

Puts the metadata key value pair for a specified schema version ID. A maximum of 10 key value pairs will be allowed per schema version. They can be added over one or more calls.

PutWorkflowRunProperties

Puts the specified workflow run properties for the given workflow run. If a property already exists for the specified run, then it overrides the value otherwise adds the property to existing properties.

QuerySchemaVersionMetadata

Queries for the schema version metadata information.

RegisterSchemaVersion

Adds a new version to the existing schema. Returns an error if new version of schema does not meet the compatibility requirements of the schema set. This API will not create a new schema set and will return a 404 error if the schema set is not already present in the Schema Registry.

RemoveSchemaVersionMetadata

Removes a key value pair from the schema version metadata for the specified schema version ID.

ResetJobBookmark

Resets a bookmark entry.

ResumeWorkflowRun

Restarts selected nodes of a previous partially completed workflow run and resumes the workflow run. The selected nodes and all nodes that are downstream from the selected nodes are run.

SearchTables

Searches a set of tables based on properties in the table metadata as well as on the parent database. You can search against text or filter conditions.

StartBlueprintRun

Starts a new run of the specified blueprint.

StartCrawler

Starts a crawl using the specified crawler, regardless

StartCrawlerSchedule

Changes the schedule state of the specified crawler to

StartExportLabelsTaskRun

Begins an asynchronous task to export all labeled data for a particular transform. This

StartImportLabelsTaskRun

Enables you to provide additional labels (examples of truth) to be used to teach the

StartJobRun

Starts a job run using a job definition.

StartMLEvaluationTaskRun

Starts a task to estimate the quality of the transform.

StartMLLabelingSetGenerationTaskRun

Starts the active learning workflow for your machine learning transform to improve the

StartTrigger

Starts an existing trigger. See Triggering

StartWorkflowRun

Starts a new run of the specified workflow.

StopCrawler

If the specified crawler is running, stops the crawl.

StopCrawlerSchedule

Sets the schedule state of the specified crawler to

StopTrigger

Stops a specified trigger.

StopWorkflowRun

Stops the execution of the specified workflow run.

TagResource

Adds tags to a resource. A tag is a label you can assign to an Amazon Web Services resource.

UntagResource

Removes tags from a resource.

UpdateBlueprint

Updates a registered blueprint.

UpdateClassifier

Modifies an existing classifier (a GrokClassifier,

UpdateColumnStatisticsForPartition

Creates or updates partition statistics of columns.

UpdateColumnStatisticsForTable

Creates or updates table statistics of columns.

UpdateConnection

Updates a connection definition in the Data Catalog.

UpdateCrawler

Updates a crawler. If a crawler is

UpdateCrawlerSchedule

Updates the schedule of a crawler using a cron expression.

UpdateDatabase

Updates an existing database definition in a Data Catalog.

UpdateDevEndpoint

Updates a specified development endpoint.

UpdateJob

Updates an existing job definition.

UpdateMLTransform

Updates an existing machine learning transform. Call this operation to tune the algorithm parameters to achieve better results.

UpdatePartition

Updates a partition.

UpdateRegistry

Updates an existing registry which is used to hold a collection of schemas. The updated properties relate to the registry, and do not modify any of the schemas within the registry.

UpdateSchema

Updates the description, compatibility setting, or version checkpoint for a schema set.

UpdateTable

Updates a metadata table in the Data Catalog.

UpdateTrigger

Updates a trigger definition.

UpdateUserDefinedFunction

Updates an existing function definition in the Data Catalog.

UpdateWorkflow

Updates an existing workflow.