Struct aws_sdk_glue::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSGlue.

This client allows ergonomic access to a AWSGlue-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_create_partition(&self) -> BatchCreatePartition<C, M, R>

Constructs a fluent builder for the BatchCreatePartition operation.

See BatchCreatePartition for more information about the operation and its arguments.

pub fn batch_delete_connection(&self) -> BatchDeleteConnection<C, M, R>

Constructs a fluent builder for the BatchDeleteConnection operation.

See BatchDeleteConnection for more information about the operation and its arguments.

pub fn batch_delete_partition(&self) -> BatchDeletePartition<C, M, R>

Constructs a fluent builder for the BatchDeletePartition operation.

See BatchDeletePartition for more information about the operation and its arguments.

pub fn batch_delete_table(&self) -> BatchDeleteTable<C, M, R>

Constructs a fluent builder for the BatchDeleteTable operation.

See BatchDeleteTable for more information about the operation and its arguments.

pub fn batch_delete_table_version(&self) -> BatchDeleteTableVersion<C, M, R>

Constructs a fluent builder for the BatchDeleteTableVersion operation.

See BatchDeleteTableVersion for more information about the operation and its arguments.

pub fn batch_get_blueprints(&self) -> BatchGetBlueprints<C, M, R>

Constructs a fluent builder for the BatchGetBlueprints operation.

See BatchGetBlueprints for more information about the operation and its arguments.

pub fn batch_get_crawlers(&self) -> BatchGetCrawlers<C, M, R>

Constructs a fluent builder for the BatchGetCrawlers operation.

See BatchGetCrawlers for more information about the operation and its arguments.

pub fn batch_get_dev_endpoints(&self) -> BatchGetDevEndpoints<C, M, R>

Constructs a fluent builder for the BatchGetDevEndpoints operation.

See BatchGetDevEndpoints for more information about the operation and its arguments.

pub fn batch_get_jobs(&self) -> BatchGetJobs<C, M, R>

Constructs a fluent builder for the BatchGetJobs operation.

See BatchGetJobs for more information about the operation and its arguments.

pub fn batch_get_partition(&self) -> BatchGetPartition<C, M, R>

Constructs a fluent builder for the BatchGetPartition operation.

See BatchGetPartition for more information about the operation and its arguments.

pub fn batch_get_triggers(&self) -> BatchGetTriggers<C, M, R>

Constructs a fluent builder for the BatchGetTriggers operation.

See BatchGetTriggers for more information about the operation and its arguments.

pub fn batch_get_workflows(&self) -> BatchGetWorkflows<C, M, R>

Constructs a fluent builder for the BatchGetWorkflows operation.

See BatchGetWorkflows for more information about the operation and its arguments.

pub fn batch_stop_job_run(&self) -> BatchStopJobRun<C, M, R>

Constructs a fluent builder for the BatchStopJobRun operation.

See BatchStopJobRun for more information about the operation and its arguments.

pub fn batch_update_partition(&self) -> BatchUpdatePartition<C, M, R>

Constructs a fluent builder for the BatchUpdatePartition operation.

See BatchUpdatePartition for more information about the operation and its arguments.

pub fn cancel_ml_task_run(&self) -> CancelMLTaskRun<C, M, R>

Constructs a fluent builder for the CancelMLTaskRun operation.

See CancelMLTaskRun for more information about the operation and its arguments.

pub fn check_schema_version_validity(
&self
) -> CheckSchemaVersionValidity<C, M, R>

Constructs a fluent builder for the CheckSchemaVersionValidity operation.

See CheckSchemaVersionValidity for more information about the operation and its arguments.

pub fn create_blueprint(&self) -> CreateBlueprint<C, M, R>

Constructs a fluent builder for the CreateBlueprint operation.

See CreateBlueprint for more information about the operation and its arguments.

pub fn create_classifier(&self) -> CreateClassifier<C, M, R>

Constructs a fluent builder for the CreateClassifier operation.

See CreateClassifier for more information about the operation and its arguments.

pub fn create_connection(&self) -> CreateConnection<C, M, R>

Constructs a fluent builder for the CreateConnection operation.

See CreateConnection for more information about the operation and its arguments.

pub fn create_crawler(&self) -> CreateCrawler<C, M, R>

Constructs a fluent builder for the CreateCrawler operation.

See CreateCrawler for more information about the operation and its arguments.

pub fn create_database(&self) -> CreateDatabase<C, M, R>

Constructs a fluent builder for the CreateDatabase operation.

See CreateDatabase for more information about the operation and its arguments.

pub fn create_dev_endpoint(&self) -> CreateDevEndpoint<C, M, R>

Constructs a fluent builder for the CreateDevEndpoint operation.

See CreateDevEndpoint for more information about the operation and its arguments.

pub fn create_job(&self) -> CreateJob<C, M, R>

Constructs a fluent builder for the CreateJob operation.

See CreateJob for more information about the operation and its arguments.

pub fn create_ml_transform(&self) -> CreateMLTransform<C, M, R>

Constructs a fluent builder for the CreateMLTransform operation.

See CreateMLTransform for more information about the operation and its arguments.

pub fn create_partition(&self) -> CreatePartition<C, M, R>

Constructs a fluent builder for the CreatePartition operation.

See CreatePartition for more information about the operation and its arguments.

pub fn create_partition_index(&self) -> CreatePartitionIndex<C, M, R>

Constructs a fluent builder for the CreatePartitionIndex operation.

See CreatePartitionIndex for more information about the operation and its arguments.

pub fn create_registry(&self) -> CreateRegistry<C, M, R>

Constructs a fluent builder for the CreateRegistry operation.

See CreateRegistry for more information about the operation and its arguments.

pub fn create_schema(&self) -> CreateSchema<C, M, R>

Constructs a fluent builder for the CreateSchema operation.

See CreateSchema for more information about the operation and its arguments.

pub fn create_script(&self) -> CreateScript<C, M, R>

Constructs a fluent builder for the CreateScript operation.

See CreateScript for more information about the operation and its arguments.

pub fn create_security_configuration(
&self
) -> CreateSecurityConfiguration<C, M, R>

Constructs a fluent builder for the CreateSecurityConfiguration operation.

See CreateSecurityConfiguration for more information about the operation and its arguments.

pub fn create_table(&self) -> CreateTable<C, M, R>

Constructs a fluent builder for the CreateTable operation.

See CreateTable for more information about the operation and its arguments.

pub fn create_trigger(&self) -> CreateTrigger<C, M, R>

Constructs a fluent builder for the CreateTrigger operation.

See CreateTrigger for more information about the operation and its arguments.

pub fn create_user_defined_function(&self) -> CreateUserDefinedFunction<C, M, R>

Constructs a fluent builder for the CreateUserDefinedFunction operation.

See CreateUserDefinedFunction for more information about the operation and its arguments.

pub fn create_workflow(&self) -> CreateWorkflow<C, M, R>

Constructs a fluent builder for the CreateWorkflow operation.

See CreateWorkflow for more information about the operation and its arguments.

pub fn delete_blueprint(&self) -> DeleteBlueprint<C, M, R>

Constructs a fluent builder for the DeleteBlueprint operation.

See DeleteBlueprint for more information about the operation and its arguments.

pub fn delete_classifier(&self) -> DeleteClassifier<C, M, R>

Constructs a fluent builder for the DeleteClassifier operation.

See DeleteClassifier for more information about the operation and its arguments.

pub fn delete_column_statistics_for_partition(
&self
) -> DeleteColumnStatisticsForPartition<C, M, R>

Constructs a fluent builder for the DeleteColumnStatisticsForPartition operation.

See DeleteColumnStatisticsForPartition for more information about the operation and its arguments.

pub fn delete_column_statistics_for_table(
&self
) -> DeleteColumnStatisticsForTable<C, M, R>

Constructs a fluent builder for the DeleteColumnStatisticsForTable operation.

See DeleteColumnStatisticsForTable for more information about the operation and its arguments.

pub fn delete_connection(&self) -> DeleteConnection<C, M, R>

Constructs a fluent builder for the DeleteConnection operation.

See DeleteConnection for more information about the operation and its arguments.

pub fn delete_crawler(&self) -> DeleteCrawler<C, M, R>

Constructs a fluent builder for the DeleteCrawler operation.

See DeleteCrawler for more information about the operation and its arguments.

pub fn delete_database(&self) -> DeleteDatabase<C, M, R>

Constructs a fluent builder for the DeleteDatabase operation.

See DeleteDatabase for more information about the operation and its arguments.

pub fn delete_dev_endpoint(&self) -> DeleteDevEndpoint<C, M, R>

Constructs a fluent builder for the DeleteDevEndpoint operation.

See DeleteDevEndpoint for more information about the operation and its arguments.

pub fn delete_job(&self) -> DeleteJob<C, M, R>

Constructs a fluent builder for the DeleteJob operation.

See DeleteJob for more information about the operation and its arguments.

pub fn delete_ml_transform(&self) -> DeleteMLTransform<C, M, R>

Constructs a fluent builder for the DeleteMLTransform operation.

See DeleteMLTransform for more information about the operation and its arguments.

pub fn delete_partition(&self) -> DeletePartition<C, M, R>

Constructs a fluent builder for the DeletePartition operation.

See DeletePartition for more information about the operation and its arguments.

pub fn delete_partition_index(&self) -> DeletePartitionIndex<C, M, R>

Constructs a fluent builder for the DeletePartitionIndex operation.

See DeletePartitionIndex for more information about the operation and its arguments.

pub fn delete_registry(&self) -> DeleteRegistry<C, M, R>

Constructs a fluent builder for the DeleteRegistry operation.

See DeleteRegistry for more information about the operation and its arguments.

pub fn delete_resource_policy(&self) -> DeleteResourcePolicy<C, M, R>

Constructs a fluent builder for the DeleteResourcePolicy operation.

See DeleteResourcePolicy for more information about the operation and its arguments.

pub fn delete_schema(&self) -> DeleteSchema<C, M, R>

Constructs a fluent builder for the DeleteSchema operation.

See DeleteSchema for more information about the operation and its arguments.

pub fn delete_schema_versions(&self) -> DeleteSchemaVersions<C, M, R>

Constructs a fluent builder for the DeleteSchemaVersions operation.

See DeleteSchemaVersions for more information about the operation and its arguments.

pub fn delete_security_configuration(
&self
) -> DeleteSecurityConfiguration<C, M, R>

Constructs a fluent builder for the DeleteSecurityConfiguration operation.

See DeleteSecurityConfiguration for more information about the operation and its arguments.

pub fn delete_table(&self) -> DeleteTable<C, M, R>

Constructs a fluent builder for the DeleteTable operation.

See DeleteTable for more information about the operation and its arguments.

pub fn delete_table_version(&self) -> DeleteTableVersion<C, M, R>

Constructs a fluent builder for the DeleteTableVersion operation.

See DeleteTableVersion for more information about the operation and its arguments.

pub fn delete_trigger(&self) -> DeleteTrigger<C, M, R>

Constructs a fluent builder for the DeleteTrigger operation.

See DeleteTrigger for more information about the operation and its arguments.

pub fn delete_user_defined_function(&self) -> DeleteUserDefinedFunction<C, M, R>

Constructs a fluent builder for the DeleteUserDefinedFunction operation.

See DeleteUserDefinedFunction for more information about the operation and its arguments.

pub fn delete_workflow(&self) -> DeleteWorkflow<C, M, R>

Constructs a fluent builder for the DeleteWorkflow operation.

See DeleteWorkflow for more information about the operation and its arguments.

pub fn get_blueprint(&self) -> GetBlueprint<C, M, R>

Constructs a fluent builder for the GetBlueprint operation.

See GetBlueprint for more information about the operation and its arguments.

pub fn get_blueprint_run(&self) -> GetBlueprintRun<C, M, R>

Constructs a fluent builder for the GetBlueprintRun operation.

See GetBlueprintRun for more information about the operation and its arguments.

pub fn get_blueprint_runs(&self) -> GetBlueprintRuns<C, M, R>

Constructs a fluent builder for the GetBlueprintRuns operation.

See GetBlueprintRuns for more information about the operation and its arguments.

pub fn get_catalog_import_status(&self) -> GetCatalogImportStatus<C, M, R>

Constructs a fluent builder for the GetCatalogImportStatus operation.

See GetCatalogImportStatus for more information about the operation and its arguments.

pub fn get_classifier(&self) -> GetClassifier<C, M, R>

Constructs a fluent builder for the GetClassifier operation.

See GetClassifier for more information about the operation and its arguments.

pub fn get_classifiers(&self) -> GetClassifiers<C, M, R>

Constructs a fluent builder for the GetClassifiers operation.

See GetClassifiers for more information about the operation and its arguments.

pub fn get_column_statistics_for_partition(
&self
) -> GetColumnStatisticsForPartition<C, M, R>

Constructs a fluent builder for the GetColumnStatisticsForPartition operation.

See GetColumnStatisticsForPartition for more information about the operation and its arguments.

pub fn get_column_statistics_for_table(
&self
) -> GetColumnStatisticsForTable<C, M, R>

Constructs a fluent builder for the GetColumnStatisticsForTable operation.

See GetColumnStatisticsForTable for more information about the operation and its arguments.

pub fn get_connection(&self) -> GetConnection<C, M, R>

Constructs a fluent builder for the GetConnection operation.

See GetConnection for more information about the operation and its arguments.

pub fn get_connections(&self) -> GetConnections<C, M, R>

Constructs a fluent builder for the GetConnections operation.

See GetConnections for more information about the operation and its arguments.

pub fn get_crawler(&self) -> GetCrawler<C, M, R>

Constructs a fluent builder for the GetCrawler operation.

See GetCrawler for more information about the operation and its arguments.

pub fn get_crawler_metrics(&self) -> GetCrawlerMetrics<C, M, R>

Constructs a fluent builder for the GetCrawlerMetrics operation.

See GetCrawlerMetrics for more information about the operation and its arguments.

pub fn get_crawlers(&self) -> GetCrawlers<C, M, R>

Constructs a fluent builder for the GetCrawlers operation.

See GetCrawlers for more information about the operation and its arguments.

pub fn get_database(&self) -> GetDatabase<C, M, R>

Constructs a fluent builder for the GetDatabase operation.

See GetDatabase for more information about the operation and its arguments.

pub fn get_databases(&self) -> GetDatabases<C, M, R>

Constructs a fluent builder for the GetDatabases operation.

See GetDatabases for more information about the operation and its arguments.

pub fn get_data_catalog_encryption_settings(
&self
) -> GetDataCatalogEncryptionSettings<C, M, R>

Constructs a fluent builder for the GetDataCatalogEncryptionSettings operation.

See GetDataCatalogEncryptionSettings for more information about the operation and its arguments.

pub fn get_dataflow_graph(&self) -> GetDataflowGraph<C, M, R>

Constructs a fluent builder for the GetDataflowGraph operation.

See GetDataflowGraph for more information about the operation and its arguments.

pub fn get_dev_endpoint(&self) -> GetDevEndpoint<C, M, R>

Constructs a fluent builder for the GetDevEndpoint operation.

See GetDevEndpoint for more information about the operation and its arguments.

pub fn get_dev_endpoints(&self) -> GetDevEndpoints<C, M, R>

Constructs a fluent builder for the GetDevEndpoints operation.

See GetDevEndpoints for more information about the operation and its arguments.

pub fn get_job(&self) -> GetJob<C, M, R>

Constructs a fluent builder for the GetJob operation.

See GetJob for more information about the operation and its arguments.

pub fn get_job_bookmark(&self) -> GetJobBookmark<C, M, R>

Constructs a fluent builder for the GetJobBookmark operation.

See GetJobBookmark for more information about the operation and its arguments.

pub fn get_job_run(&self) -> GetJobRun<C, M, R>

Constructs a fluent builder for the GetJobRun operation.

See GetJobRun for more information about the operation and its arguments.

pub fn get_job_runs(&self) -> GetJobRuns<C, M, R>

Constructs a fluent builder for the GetJobRuns operation.

See GetJobRuns for more information about the operation and its arguments.

pub fn get_jobs(&self) -> GetJobs<C, M, R>

Constructs a fluent builder for the GetJobs operation.

See GetJobs for more information about the operation and its arguments.

pub fn get_mapping(&self) -> GetMapping<C, M, R>

Constructs a fluent builder for the GetMapping operation.

See GetMapping for more information about the operation and its arguments.

pub fn get_ml_task_run(&self) -> GetMLTaskRun<C, M, R>

Constructs a fluent builder for the GetMLTaskRun operation.

See GetMLTaskRun for more information about the operation and its arguments.

pub fn get_ml_task_runs(&self) -> GetMLTaskRuns<C, M, R>

Constructs a fluent builder for the GetMLTaskRuns operation.

See GetMLTaskRuns for more information about the operation and its arguments.

pub fn get_ml_transform(&self) -> GetMLTransform<C, M, R>

Constructs a fluent builder for the GetMLTransform operation.

See GetMLTransform for more information about the operation and its arguments.

pub fn get_ml_transforms(&self) -> GetMLTransforms<C, M, R>

Constructs a fluent builder for the GetMLTransforms operation.

See GetMLTransforms for more information about the operation and its arguments.

pub fn get_partition(&self) -> GetPartition<C, M, R>

Constructs a fluent builder for the GetPartition operation.

See GetPartition for more information about the operation and its arguments.

pub fn get_partition_indexes(&self) -> GetPartitionIndexes<C, M, R>

Constructs a fluent builder for the GetPartitionIndexes operation.

See GetPartitionIndexes for more information about the operation and its arguments.

pub fn get_partitions(&self) -> GetPartitions<C, M, R>

Constructs a fluent builder for the GetPartitions operation.

See GetPartitions for more information about the operation and its arguments.

pub fn get_plan(&self) -> GetPlan<C, M, R>

Constructs a fluent builder for the GetPlan operation.

See GetPlan for more information about the operation and its arguments.

pub fn get_registry(&self) -> GetRegistry<C, M, R>

Constructs a fluent builder for the GetRegistry operation.

See GetRegistry for more information about the operation and its arguments.

pub fn get_resource_policies(&self) -> GetResourcePolicies<C, M, R>

Constructs a fluent builder for the GetResourcePolicies operation.

See GetResourcePolicies for more information about the operation and its arguments.

pub fn get_resource_policy(&self) -> GetResourcePolicy<C, M, R>

Constructs a fluent builder for the GetResourcePolicy operation.

See GetResourcePolicy for more information about the operation and its arguments.

pub fn get_schema(&self) -> GetSchema<C, M, R>

Constructs a fluent builder for the GetSchema operation.

See GetSchema for more information about the operation and its arguments.

pub fn get_schema_by_definition(&self) -> GetSchemaByDefinition<C, M, R>

Constructs a fluent builder for the GetSchemaByDefinition operation.

See GetSchemaByDefinition for more information about the operation and its arguments.

pub fn get_schema_version(&self) -> GetSchemaVersion<C, M, R>

Constructs a fluent builder for the GetSchemaVersion operation.

See GetSchemaVersion for more information about the operation and its arguments.

pub fn get_schema_versions_diff(&self) -> GetSchemaVersionsDiff<C, M, R>

Constructs a fluent builder for the GetSchemaVersionsDiff operation.

See GetSchemaVersionsDiff for more information about the operation and its arguments.

pub fn get_security_configuration(&self) -> GetSecurityConfiguration<C, M, R>

Constructs a fluent builder for the GetSecurityConfiguration operation.

See GetSecurityConfiguration for more information about the operation and its arguments.

pub fn get_security_configurations(&self) -> GetSecurityConfigurations<C, M, R>

Constructs a fluent builder for the GetSecurityConfigurations operation.

See GetSecurityConfigurations for more information about the operation and its arguments.

pub fn get_table(&self) -> GetTable<C, M, R>

Constructs a fluent builder for the GetTable operation.

See GetTable for more information about the operation and its arguments.

pub fn get_tables(&self) -> GetTables<C, M, R>

Constructs a fluent builder for the GetTables operation.

See GetTables for more information about the operation and its arguments.

pub fn get_table_version(&self) -> GetTableVersion<C, M, R>

Constructs a fluent builder for the GetTableVersion operation.

See GetTableVersion for more information about the operation and its arguments.

pub fn get_table_versions(&self) -> GetTableVersions<C, M, R>

Constructs a fluent builder for the GetTableVersions operation.

See GetTableVersions for more information about the operation and its arguments.

pub fn get_tags(&self) -> GetTags<C, M, R>

Constructs a fluent builder for the GetTags operation.

See GetTags for more information about the operation and its arguments.

pub fn get_trigger(&self) -> GetTrigger<C, M, R>

Constructs a fluent builder for the GetTrigger operation.

See GetTrigger for more information about the operation and its arguments.

pub fn get_triggers(&self) -> GetTriggers<C, M, R>

Constructs a fluent builder for the GetTriggers operation.

See GetTriggers for more information about the operation and its arguments.

pub fn get_user_defined_function(&self) -> GetUserDefinedFunction<C, M, R>

Constructs a fluent builder for the GetUserDefinedFunction operation.

See GetUserDefinedFunction for more information about the operation and its arguments.

pub fn get_user_defined_functions(&self) -> GetUserDefinedFunctions<C, M, R>

Constructs a fluent builder for the GetUserDefinedFunctions operation.

See GetUserDefinedFunctions for more information about the operation and its arguments.

pub fn get_workflow(&self) -> GetWorkflow<C, M, R>

Constructs a fluent builder for the GetWorkflow operation.

See GetWorkflow for more information about the operation and its arguments.

pub fn get_workflow_run(&self) -> GetWorkflowRun<C, M, R>

Constructs a fluent builder for the GetWorkflowRun operation.

See GetWorkflowRun for more information about the operation and its arguments.

pub fn get_workflow_run_properties(&self) -> GetWorkflowRunProperties<C, M, R>

Constructs a fluent builder for the GetWorkflowRunProperties operation.

See GetWorkflowRunProperties for more information about the operation and its arguments.

pub fn get_workflow_runs(&self) -> GetWorkflowRuns<C, M, R>

Constructs a fluent builder for the GetWorkflowRuns operation.

See GetWorkflowRuns for more information about the operation and its arguments.

pub fn import_catalog_to_glue(&self) -> ImportCatalogToGlue<C, M, R>

Constructs a fluent builder for the ImportCatalogToGlue operation.

See ImportCatalogToGlue for more information about the operation and its arguments.

pub fn list_blueprints(&self) -> ListBlueprints<C, M, R>

Constructs a fluent builder for the ListBlueprints operation.

See ListBlueprints for more information about the operation and its arguments.

pub fn list_crawlers(&self) -> ListCrawlers<C, M, R>

Constructs a fluent builder for the ListCrawlers operation.

See ListCrawlers for more information about the operation and its arguments.

pub fn list_dev_endpoints(&self) -> ListDevEndpoints<C, M, R>

Constructs a fluent builder for the ListDevEndpoints operation.

See ListDevEndpoints for more information about the operation and its arguments.

pub fn list_jobs(&self) -> ListJobs<C, M, R>

Constructs a fluent builder for the ListJobs operation.

See ListJobs for more information about the operation and its arguments.

pub fn list_ml_transforms(&self) -> ListMLTransforms<C, M, R>

Constructs a fluent builder for the ListMLTransforms operation.

See ListMLTransforms for more information about the operation and its arguments.

pub fn list_registries(&self) -> ListRegistries<C, M, R>

Constructs a fluent builder for the ListRegistries operation.

See ListRegistries for more information about the operation and its arguments.

pub fn list_schemas(&self) -> ListSchemas<C, M, R>

Constructs a fluent builder for the ListSchemas operation.

See ListSchemas for more information about the operation and its arguments.

pub fn list_schema_versions(&self) -> ListSchemaVersions<C, M, R>

Constructs a fluent builder for the ListSchemaVersions operation.

See ListSchemaVersions for more information about the operation and its arguments.

pub fn list_triggers(&self) -> ListTriggers<C, M, R>

Constructs a fluent builder for the ListTriggers operation.

See ListTriggers for more information about the operation and its arguments.

pub fn list_workflows(&self) -> ListWorkflows<C, M, R>

Constructs a fluent builder for the ListWorkflows operation.

See ListWorkflows for more information about the operation and its arguments.

pub fn put_data_catalog_encryption_settings(
&self
) -> PutDataCatalogEncryptionSettings<C, M, R>

Constructs a fluent builder for the PutDataCatalogEncryptionSettings operation.

See PutDataCatalogEncryptionSettings for more information about the operation and its arguments.

pub fn put_resource_policy(&self) -> PutResourcePolicy<C, M, R>

Constructs a fluent builder for the PutResourcePolicy operation.

See PutResourcePolicy for more information about the operation and its arguments.

pub fn put_schema_version_metadata(&self) -> PutSchemaVersionMetadata<C, M, R>

Constructs a fluent builder for the PutSchemaVersionMetadata operation.

See PutSchemaVersionMetadata for more information about the operation and its arguments.

pub fn put_workflow_run_properties(&self) -> PutWorkflowRunProperties<C, M, R>

Constructs a fluent builder for the PutWorkflowRunProperties operation.

See PutWorkflowRunProperties for more information about the operation and its arguments.

pub fn query_schema_version_metadata(
&self
) -> QuerySchemaVersionMetadata<C, M, R>

Constructs a fluent builder for the QuerySchemaVersionMetadata operation.

See QuerySchemaVersionMetadata for more information about the operation and its arguments.

pub fn register_schema_version(&self) -> RegisterSchemaVersion<C, M, R>

Constructs a fluent builder for the RegisterSchemaVersion operation.

See RegisterSchemaVersion for more information about the operation and its arguments.

pub fn remove_schema_version_metadata(
&self
) -> RemoveSchemaVersionMetadata<C, M, R>

Constructs a fluent builder for the RemoveSchemaVersionMetadata operation.

See RemoveSchemaVersionMetadata for more information about the operation and its arguments.

pub fn reset_job_bookmark(&self) -> ResetJobBookmark<C, M, R>

Constructs a fluent builder for the ResetJobBookmark operation.

See ResetJobBookmark for more information about the operation and its arguments.

pub fn resume_workflow_run(&self) -> ResumeWorkflowRun<C, M, R>

Constructs a fluent builder for the ResumeWorkflowRun operation.

See ResumeWorkflowRun for more information about the operation and its arguments.

pub fn search_tables(&self) -> SearchTables<C, M, R>

Constructs a fluent builder for the SearchTables operation.

See SearchTables for more information about the operation and its arguments.

pub fn start_blueprint_run(&self) -> StartBlueprintRun<C, M, R>

Constructs a fluent builder for the StartBlueprintRun operation.

See StartBlueprintRun for more information about the operation and its arguments.

pub fn start_crawler(&self) -> StartCrawler<C, M, R>

Constructs a fluent builder for the StartCrawler operation.

See StartCrawler for more information about the operation and its arguments.

pub fn start_crawler_schedule(&self) -> StartCrawlerSchedule<C, M, R>

Constructs a fluent builder for the StartCrawlerSchedule operation.

See StartCrawlerSchedule for more information about the operation and its arguments.

pub fn start_export_labels_task_run(&self) -> StartExportLabelsTaskRun<C, M, R>

Constructs a fluent builder for the StartExportLabelsTaskRun operation.

See StartExportLabelsTaskRun for more information about the operation and its arguments.

pub fn start_import_labels_task_run(&self) -> StartImportLabelsTaskRun<C, M, R>

Constructs a fluent builder for the StartImportLabelsTaskRun operation.

See StartImportLabelsTaskRun for more information about the operation and its arguments.

pub fn start_job_run(&self) -> StartJobRun<C, M, R>

Constructs a fluent builder for the StartJobRun operation.

See StartJobRun for more information about the operation and its arguments.

pub fn start_ml_evaluation_task_run(&self) -> StartMLEvaluationTaskRun<C, M, R>

Constructs a fluent builder for the StartMLEvaluationTaskRun operation.

See StartMLEvaluationTaskRun for more information about the operation and its arguments.

pub fn start_ml_labeling_set_generation_task_run(
&self
) -> StartMLLabelingSetGenerationTaskRun<C, M, R>

Constructs a fluent builder for the StartMLLabelingSetGenerationTaskRun operation.

See StartMLLabelingSetGenerationTaskRun for more information about the operation and its arguments.

pub fn start_trigger(&self) -> StartTrigger<C, M, R>

Constructs a fluent builder for the StartTrigger operation.

See StartTrigger for more information about the operation and its arguments.

pub fn start_workflow_run(&self) -> StartWorkflowRun<C, M, R>

Constructs a fluent builder for the StartWorkflowRun operation.

See StartWorkflowRun for more information about the operation and its arguments.

pub fn stop_crawler(&self) -> StopCrawler<C, M, R>

Constructs a fluent builder for the StopCrawler operation.

See StopCrawler for more information about the operation and its arguments.

pub fn stop_crawler_schedule(&self) -> StopCrawlerSchedule<C, M, R>

Constructs a fluent builder for the StopCrawlerSchedule operation.

See StopCrawlerSchedule for more information about the operation and its arguments.

pub fn stop_trigger(&self) -> StopTrigger<C, M, R>

Constructs a fluent builder for the StopTrigger operation.

See StopTrigger for more information about the operation and its arguments.

pub fn stop_workflow_run(&self) -> StopWorkflowRun<C, M, R>

Constructs a fluent builder for the StopWorkflowRun operation.

See StopWorkflowRun for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_blueprint(&self) -> UpdateBlueprint<C, M, R>

Constructs a fluent builder for the UpdateBlueprint operation.

See UpdateBlueprint for more information about the operation and its arguments.

pub fn update_classifier(&self) -> UpdateClassifier<C, M, R>

Constructs a fluent builder for the UpdateClassifier operation.

See UpdateClassifier for more information about the operation and its arguments.

pub fn update_column_statistics_for_partition(
&self
) -> UpdateColumnStatisticsForPartition<C, M, R>

Constructs a fluent builder for the UpdateColumnStatisticsForPartition operation.

See UpdateColumnStatisticsForPartition for more information about the operation and its arguments.

pub fn update_column_statistics_for_table(
&self
) -> UpdateColumnStatisticsForTable<C, M, R>

Constructs a fluent builder for the UpdateColumnStatisticsForTable operation.

See UpdateColumnStatisticsForTable for more information about the operation and its arguments.

pub fn update_connection(&self) -> UpdateConnection<C, M, R>

Constructs a fluent builder for the UpdateConnection operation.

See UpdateConnection for more information about the operation and its arguments.

pub fn update_crawler(&self) -> UpdateCrawler<C, M, R>

Constructs a fluent builder for the UpdateCrawler operation.

See UpdateCrawler for more information about the operation and its arguments.

pub fn update_crawler_schedule(&self) -> UpdateCrawlerSchedule<C, M, R>

Constructs a fluent builder for the UpdateCrawlerSchedule operation.

See UpdateCrawlerSchedule for more information about the operation and its arguments.

pub fn update_database(&self) -> UpdateDatabase<C, M, R>

Constructs a fluent builder for the UpdateDatabase operation.

See UpdateDatabase for more information about the operation and its arguments.

pub fn update_dev_endpoint(&self) -> UpdateDevEndpoint<C, M, R>

Constructs a fluent builder for the UpdateDevEndpoint operation.

See UpdateDevEndpoint for more information about the operation and its arguments.

pub fn update_job(&self) -> UpdateJob<C, M, R>

Constructs a fluent builder for the UpdateJob operation.

See UpdateJob for more information about the operation and its arguments.

pub fn update_ml_transform(&self) -> UpdateMLTransform<C, M, R>

Constructs a fluent builder for the UpdateMLTransform operation.

See UpdateMLTransform for more information about the operation and its arguments.

pub fn update_partition(&self) -> UpdatePartition<C, M, R>

Constructs a fluent builder for the UpdatePartition operation.

See UpdatePartition for more information about the operation and its arguments.

pub fn update_registry(&self) -> UpdateRegistry<C, M, R>

Constructs a fluent builder for the UpdateRegistry operation.

See UpdateRegistry for more information about the operation and its arguments.

pub fn update_schema(&self) -> UpdateSchema<C, M, R>

Constructs a fluent builder for the UpdateSchema operation.

See UpdateSchema for more information about the operation and its arguments.

pub fn update_table(&self) -> UpdateTable<C, M, R>

Constructs a fluent builder for the UpdateTable operation.

See UpdateTable for more information about the operation and its arguments.

pub fn update_trigger(&self) -> UpdateTrigger<C, M, R>

Constructs a fluent builder for the UpdateTrigger operation.

See UpdateTrigger for more information about the operation and its arguments.

pub fn update_user_defined_function(&self) -> UpdateUserDefinedFunction<C, M, R>

Constructs a fluent builder for the UpdateUserDefinedFunction operation.

See UpdateUserDefinedFunction for more information about the operation and its arguments.

pub fn update_workflow(&self) -> UpdateWorkflow<C, M, R>

Constructs a fluent builder for the UpdateWorkflow operation.

See UpdateWorkflow for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.